% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bridge_net.R
\name{as_bridge_net}
\alias{as_bridge_net}
\alias{as_bridge_net.bridge_net}
\alias{as_bridge_net.igraph}
\alias{as_bridge_net.network}
\alias{as_bridge_net.tbl_graph}
\alias{print.bridge_net}
\title{A data structure that effectively bridges \code{igraph} and \code{network} objects.}
\usage{
as_bridge_net(x)

\method{as_bridge_net}{bridge_net}(x)

\method{as_bridge_net}{igraph}(x)

\method{as_bridge_net}{network}(x)

\method{as_bridge_net}{tbl_graph}(x)

\method{print}{bridge_net}(x, .nrow = 3L)
}
\arguments{
\item{x}{\code{igraph} (\code{\link[igraph:graph]{igraph::graph}}), \code{\link[network:network]{network::network}}, or
\code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}}.}
}
\description{
A data structure that effectively bridges \code{igraph} and \code{network} objects.
}
