% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_network.R
\docType{import}
\name{as_network}
\alias{as_network}
\alias{as_network.bridge_net}
\alias{as_network.igraph}
\alias{as_network.network}
\alias{as_network.tbl_graph}
\alias{as.network.igraph}
\alias{reexports}
\alias{as.network}
\title{Conversion to \code{network} objects.}
\usage{
as_network(x, ...)

\method{as_network}{bridge_net}(x, .actor = TRUE, ...)

\method{as_network}{igraph}(x, .actor = TRUE, ...)

\method{as_network}{network}(x, ...)

\method{as_network}{tbl_graph}(x)

\method{as.network}{igraph}(x, ...)
}
\arguments{
\item{x}{\code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) or \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} object.}
}
\value{
A \code{\link[network:network]{network::network}} object.
}
\seealso{
\code{\link[=as_igraph]{as_igraph()}}, \code{\link[intergraph:asNetwork]{intergraph::asNetwork()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{network}{\code{\link[network]{as.network}}}
}}

