% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{bip_clarify_actors}
\alias{bip_clarify_actors}
\title{Assign vertex attributes to the distinguish sets in \code{network} objects.}
\usage{
bip_clarify_actors(x)
}
\arguments{
\item{x}{A \code{\link[network:network]{network::network}} object.}
}
\value{
A \code{\link[network:network]{network::network}} object with a new, \code{logical} vertex attribute named
\code{is_actor}.
}
\description{
Assign vertex attributes to the distinguish sets in \code{network} objects.
}
\examples{
library(snatools)

sw_matrix <- matrix(
  c(1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0,
    1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 
    1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 
    0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1,
    1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 
    1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1,
    1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 
    0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0,
    0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 
    1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0),
  nrow = 14L, ncol = 18L,
  dimnames = list(
  c("E1", "E2", "E3", "E4", "E5", "E6", "E7",
    "E8", "E9", "E10", "E11", "E12", "E13", "E14"),
  c("EVELYN", "LAURA", "THERESA", "BRENDA", "CHARLOTTE", "FRANCES",
    "ELEANOR", "PEARL", "RUTH", "VERNE", "MYRNA", "KATHERINE", 
    "SYLVIA", "NORA", "HELEN", "DOROTHY", "OLIVIA", "FLORA"))
    )
    
(southern_women_nw <- sw_matrix \%>\% 
  t() \%>\% 
  network::as.network.matrix())
  
southern_women_nw \%>\% vrt_get_attrs()
  
(actors_clarified <- bip_clarify_actors(southern_women_nw))

actors_clarified \%>\% vrt_get_attrs()

}
