% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{bip_swap_modes}
\alias{bip_swap_modes}
\alias{bip_swap_modes.igraph}
\alias{bip_swap_modes.network}
\title{Swap vertex modes in bipartite graphs.}
\usage{
bip_swap_modes(x, ...)

\method{bip_swap_modes}{igraph}(x)

\method{bip_swap_modes}{network}(x, clarify_actors = TRUE)
}
\arguments{
\item{x}{A bipartite \code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) or \code{\link[network:network]{network::network}} object.}

\item{clarify_actors}{Whether to also run \code{bip_clarify_actors()} on \code{\link[network:network]{network::network}}
objects.}
}
\description{
Swap vertex modes in bipartite graphs.
}
\examples{
library(snatools)

sw_matrix <- matrix(
  c(1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0,
    1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 
    1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 
    0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1,
    1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 
    1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1,
    1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 
    0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0,
    0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 
    1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0),
  nrow = 14L, ncol = 18L,
  dimnames = list(
  c("E1", "E2", "E3", "E4", "E5", "E6", "E7",
    "E8", "E9", "E10", "E11", "E12", "E13", "E14"),
  c("EVELYN", "LAURA", "THERESA", "BRENDA", "CHARLOTTE", "FRANCES",
    "ELEANOR", "PEARL", "RUTH", "VERNE", "MYRNA", "KATHERINE", 
    "SYLVIA", "NORA", "HELEN", "DOROTHY", "OLIVIA", "FLORA"))
    )
    
(southern_women_ig <- sw_matrix \%>\% 
  t() \%>\% 
  igraph::graph_from_incidence_matrix())

southern_women_ig \%>\% vrt_get_attrs()

southern_women_ig \%>\% 
  bip_swap_modes() \%>\%
  vrt_get_attrs()

(southern_women_nw <- sw_matrix \%>\% 
  network::as.network.matrix() \%>\% 
  bip_clarify_actors())
  
southern_women_nw \%>\% vrt_get_attrs()
  
southern_women_nw \%>\%
  bip_swap_modes() \%>\%
  vrt_get_attrs()

}
