% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-metrics.R
\name{calc_net_density}
\alias{calc_net_density}
\title{Calculate a graph object's density.}
\usage{
calc_net_density(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, or\code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{double} indicating the ratio of existing edges to maximum potential edges.
}
\description{
Calculate a graph object's density.
}
\examples{
library(snatools)

ig <- c(0, 1, 0, 1,
        1, 0, 0, 0,
        0, 1, 0, 1,
        1, 0, 1, 0) \%>\% 
  matrix(nrow = 4, byrow = TRUE) \%>\% 
  igraph::graph_from_adjacency_matrix()
  
nw <- ig \%>\% 
  as_network()
 
ig \%>\% 
  calc_net_density()
  
ig \%>\% 
  net_as_undirected() \%>\% 
  calc_net_density()

nw \%>\% 
  calc_net_density()

nw \%>\% 
  net_as_undirected() \%>\% 
  calc_net_density()

}
\seealso{
\code{\link[igraph:edge_density]{igraph::edge_density()}}, \code{\link[sna:gden]{sna::gden()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
