% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_graph.R
\name{clean_graph}
\alias{clean_graph}
\alias{clean_graph.igraph}
\alias{clean_graph.network}
\title{Clean a graph object's internals and standardize attributes.}
\usage{
clean_graph(x, ...)

\method{clean_graph}{igraph}(x, actor_type = TRUE)

\method{clean_graph}{network}(x, ...)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A cleaned version of \code{x} with safe attributes.
}
\description{
Clean a graph object's internals and standardize attributes.
}
\examples{
library(snatools)

data("Koenigsberg", package = "igraphdata")

Koenigsberg \%>\% vrt_attrs()

Koenigsberg \%>\% clean_graph()


}
\seealso{
\code{\link[=as_igraph]{as_igraph()}}, \code{\link[=as_network]{as_network()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
