% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{davis_southern_women}
\alias{davis_southern_women}
\title{Davis' Southern Women}
\format{\code{network_dataset}}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm}
}
\usage{
davis_southern_women
}
\description{
Davis' Southern Women
}
\details{
\strong{Background}, adapted from \url{http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm}: \cr
\itemize{
\item In the 1950s John Gagnon collected sociometric choice data from 67 prison inmates. All
were asked, "What fellows on the tier are you closest friends with?" Each was free to
choose as few or as many "friends" as he desired. The data were analyzed by MacRae and
characterized by him as "less clear cut" in their internal structure than similar data
from schools or residential populations.
}
}
\examples{
data("davis_southern_women", package = "snatools")

davis_southern_women

as_igraph(davis_southern_women)

as_network(davis_southern_women)

}
\references{
\itemize{
\item Breiger R. (1974). The duality of persons and groups. Social Forces, 53, 181-190.
\item Davis, A et al. (1941). Deep South. Chicago: University of Chicago Press.
}
}
\keyword{datasets}
