% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defaulters}
\alias{defaulters}
\alias{\%||\%}
\alias{\%{}\%}
\alias{\%{NA}\%}
\alias{\%{T}\%}
\alias{\%{F}\%}
\title{Conditionally replace objects with default values.}
\usage{
.lhs \%||\% .rhs

.lhs \%{}\% .rhs

.lhs \%{NA}\% .rhs

.lhs \%{T}\% .rhs

.lhs \%{F}\% .rhs
}
\arguments{
\item{.lhs}{Object to test.}

\item{.rhs}{Object to return if test returns \code{FALSE} for \code{.lhs}.}
}
\value{
\code{.lhs} or \code{.rhs}
}
\description{
Conditionally replace objects with default values.
}
\details{
\itemize{
\item If \code{.lhs} is...
\itemize{
\item \code{\%||\%}: \code{NULL}
\item \code{\%{}\%}: empty (\code{length(.lhs) == 0L})
\item \code{\%{NA}\%}: \code{NA}
\item \code{\%{T}\%}: \code{TRUE}
\item \code{\%{F}\%}: \code{FALSE}
}
\item ... returns \code{.rhs}.
\itemize{
\item Returns \code{.lhs} otherwise.
}
}
}
\examples{
NULL \%||\% 1
1 \%||\% NULL

integer(length = 0L) \%{}\% NA
1 \%{}\% NA

NA \%{NA}\% 1
1 \%{NA}\% NA

TRUE \%{T}\% FALSE
FALSE \%{T}\% TRUE

TRUE \%{F}\% FALSE
FALSE \%{F}\% TRUE
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
