% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelists.R
\name{.fetch_edgelist}
\alias{.fetch_edgelist}
\title{Get the edge list representation of a graph as a matrix.}
\usage{
.fetch_edgelist(.g)
}
\description{
Get the edge list representation of a graph as a matrix.
}
