% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{edg_as_df}
\alias{edg_as_df}
\alias{edg_as_df.igraph}
\alias{edg_as_df.network}
\title{Extract edge attributes as a data frame.}
\usage{
edg_as_df(x)

\method{edg_as_df}{igraph}(x)

\method{edg_as_df}{network}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\description{
Extract edge attributes as a data frame.
}
