% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edg_.R
\name{edg_attr_names}
\alias{edg_attr_names}
\alias{edg_attr_names.igraph}
\alias{edg_attr_names.network}
\title{List edge attribute names.}
\usage{
edg_attr_names(x, exclude_na = TRUE)

\method{edg_attr_names}{igraph}(x, exclude_na = NULL)

\method{edg_attr_names}{network}(x, exclude_na = TRUE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{exclude_na}{\code{network} objects include a \code{"na"} attribute indicating whether
a vertex is "missing" from the network. \code{exclude_na} determines whether or not to include
this \code{"na"} attribute in returned names. \cr
\itemize{
\item Default: \code{TRUE}
\item If \code{exclude_na} is provided when \code{x} is an \code{igraph} object, it is ignored with a \code{warning}.
}}
}
\value{
\code{character} \code{vector}of \code{x}'s edge attribute names.
}
\description{
List edge attribute names.
}
\seealso{
\code{\link[igraph:edge_attr_names]{igraph::edge_attr_names()}}, \code{\link[network:list.edge.attributes]{network::list.edge.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
