% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edg_.R
\name{edg_attrs}
\alias{edg_attrs}
\alias{edg_attrs.igraph}
\alias{edg_attrs.network}
\title{Extract edge attributes.}
\usage{
edg_attrs(x, exclude_na = TRUE)

\method{edg_attrs}{igraph}(x, exclude_na = NULL)

\method{edg_attrs}{network}(x, exclude_na = TRUE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s edge attributes.
}
\description{
Extract edge attributes.
}
