% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_delete_attr}
\alias{edg_delete_attr}
\alias{edg_delete_attr.bridge_net}
\alias{edg_delete_attr.igraph}
\alias{edg_delete_attr.network}
\alias{edg_delete_attr.tbl_graph}
\title{Delete Edge Attributes}
\usage{
edg_delete_attr(x, ...)

\method{edg_delete_attr}{bridge_net}(x, ...)

\method{edg_delete_attr}{igraph}(x, ...)

\method{edg_delete_attr}{network}(x, ...)

\method{edg_delete_attr}{tbl_graph}(x, ...)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{...}{\code{character} including the name(s) of one or more edge attributes to delete
from \code{x}.}
}
\value{
\code{x}, but with \code{...} deleted.
}
\description{
Delete Edge Attributes
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

sampson_monastery \%>\% 
  edg_delete_attr("relation", "time")
  
ig \%>\% 
  edg_delete_attr("time")

nw \%>\% 
  edg_delete_attr("relation", "weight")

tidy_g \%>\% 
  edg_delete_attr("positive_relation")

}
\seealso{
\code{\link[igraph:delete_edge_attr]{igraph::delete_edge_attr()}}

\code{\link[network:delete.vertex.attribute]{network::delete.vertex.attribute()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
