% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_filter}
\alias{edg_filter}
\alias{edg_filter_se}
\alias{edg_filter.bridge_net}
\alias{edg_filter_se.bridge_net}
\alias{edg_filter.igraph}
\alias{edg_filter_se.igraph}
\alias{edg_filter.network}
\alias{edg_filter_se.network}
\title{Filter edges by their attributes.}
\usage{
edg_filter(x, ...)

edg_filter_se(x, edg_attr, .f, edg_val = NULL)

\method{edg_filter}{bridge_net}(x, ...)

\method{edg_filter_se}{bridge_net}(x, edg_attr, .f = NULL,
  edg_val = NULL)

\method{edg_filter}{igraph}(x, ...)

\method{edg_filter_se}{igraph}(x, edg_attr, .f = NULL, edg_val = NULL)

\method{edg_filter}{network}(x, ...)

\method{edg_filter_se}{network}(x, edg_attr, .f = NULL, edg_val = NULL)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{...}{\code{logical} conditions used to subset rows when said conditions evaluate to
\code{TRUE}.}

\item{edg_attr}{\code{character} specifying edge attribute to target. \code{edg_filter_se()}
only.}

\item{.f}{\code{function} (default: \code{NULL}) specifying the predicate function to apply to
\code{edg_attr}. \code{edg_filter_se()} only.}

\item{edg_val}{(default: \code{NULL}) Value to use as second argument to \code{.f} (if
applicable).}
}
\value{
\code{x}, but with edges filtered by \code{...} for \code{edg_filter()} or \code{edge_attr}, \code{.f},
and/or \code{edg_val} for \code{edg_filter_se()}.
}
\description{
Filter edges by their attributes.
}
\details{
\code{edg_filter()} uses non-standard evaluation to make syntax as convenient as possible,
while \code{edg_filter_se()} uses standard evaluation if preferred (or if \code{edg_filter()}
causes trouble).
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network() 

sampson_monastery \%>\% 
  edg_filter(relation == "liking", time == 1)

sampson_monastery \%>\% 
  edg_filter_se("relation", `==`, "liking") \%>\% 
  edg_filter_se("time", `==`, 1)

ig \%>\% 
  edg_filter(positive_relation, weight == 3)
  
ig \%>\% 
  edg_filter_se("positive_relation") \%>\% 
  edg_filter_se("weight", `==`, 3)

nw \%>\% 
  edg_filter(relation \%in\% c("blame", "disesteem"))
  
nw \%>\% 
  edg_filter_se("relation", `\%in\%`, c("blame", "disesteem"))

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
