% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_get_attr_df}
\alias{edg_get_attr_df}
\alias{edg_get_attr_df.sna_net}
\alias{edg_get_attr_df.igraph}
\alias{edg_get_attr_df.network}
\alias{edg_get_attr_df.tbl_graph}
\title{Extract all edge attributes as a data frame.}
\usage{
edg_get_attr_df(x, include_dyad = TRUE, leave_raw = TRUE)

\method{edg_get_attr_df}{sna_net}(x, include_dyad = TRUE,
  leave_raw = FALSE)

\method{edg_get_attr_df}{igraph}(x, include_dyad = TRUE,
  leave_raw = FALSE)

\method{edg_get_attr_df}{network}(x, include_dyad = TRUE,
  leave_raw = FALSE)

\method{edg_get_attr_df}{tbl_graph}(x, include_dyad = TRUE,
  leave_raw = FALSE)
}
\arguments{
\item{x}{An \code{sna_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{include_dyad}{\code{logical} (default: \code{TRUE}) indicating whether to return the \code{.ego}
and \code{.alter} of each edge's dyad.}

\item{leave_raw}{\code{logical} (default: \code{FALSE}) indicating whether to return a bare
\code{data.frame} (\code{stringsAsFactors = FALSE}) instead of an \code{sna_edges_df}.}
}
\value{
\code{sna_edges_df} \code{data.frame} with all edge attributes contained in \code{x}.
}
\description{
Extract all edge attributes as a data frame.
}
\examples{
library(snatools)

adj_mat <- rbinom(16L, 1L, 0.3) \%>\% 
  matrix(nrow = 4L, dimnames = list(letters[1:4], letters[1:4]))

ig <- adj_mat \%>\% 
  igraph::graph_from_adjacency_matrix() \%>\% 
  igraph::set_edge_attr("an_edge_attribute", value = sample(seq_len(igraph::ecount(.)))) 

edg_get_attr_df(ig)
edg_get_attr_df(ig, include_dyad = FALSE)
edg_get_attr_df(ig, leave_raw = TRUE)

nw <- network::as.network(adj_mat)
network::set.edge.attribute(nw, "an_edge_attribute", value = sample(seq_len(network::network.edgecount(nw))))

edg_get_attr_df(nw)
edg_get_attr_df(nw, include_dyad = FALSE)
edg_get_attr_df(nw, leave_raw = TRUE)

tbl_g <- tidygraph::as_tbl_graph(ig)

edg_get_attr_df(tbl_g)
edg_get_attr_df(tbl_g, include_dyad = FALSE)
edg_get_attr_df(tbl_g, leave_raw = TRUE)

sna_g <- as_sna_net(nw)

edg_get_attr_df(sna_g)

}
\seealso{
\code{\link[=edg_get_attr]{edg_get_attr()}}, \code{\link[=edg_attr_names]{edg_attr_names()}}, \code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
