% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edg_get.R
\name{edg_get_attrs}
\alias{edg_get_attrs}
\alias{edg_get_attrs.igraph}
\alias{edg_get_attrs.network}
\title{Extract edge attributes.}
\usage{
edg_get_attrs(x)

\method{edg_get_attrs}{igraph}(ig)

\method{edg_get_attrs}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s edge attributes.
}
\description{
Extract edge attributes.
}
