% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{edg_subset}
\alias{edg_subset}
\title{Subset by Edges}
\usage{
edg_subset(x, .predicate, drop_isolates = FALSE)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{.predicate}{A single predicate function describing how \code{x}'s edges should be
subset.}

\item{drop_isolates}{\code{logical} (Default: \code{FALSE}) indicating whether isolated vertices shuold be dropped.}
}
\description{
Subset by Edges
}
\examples{
library(snatools)

florence \%>\% 
  edg_subset(relation == "business")

crisis_in_cloister \%>\% 
  edg_subset(time == 1)
  
crisis_in_cloister \%>\% 
  as_network() \%>\% 
  edg_subset(time == 3 & affect < 0)
  
crisis_in_cloister \%>\% 
  edg_subset(time == 3 & affect < 0, drop_isolates = TRUE)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
