% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_to_df}
\alias{edg_to_df}
\alias{edg_to_df.bridge_net}
\alias{edg_to_df.igraph}
\alias{edg_to_df.network}
\alias{edg_to_df.tbl_graph}
\alias{print.edge_data_frame}
\title{Extract all edge attributes as a data frame.}
\usage{
edg_to_df(x, include_dyad = TRUE, leave_raw = TRUE)

\method{edg_to_df}{bridge_net}(x, include_dyad = TRUE,
  leave_raw = FALSE)

\method{edg_to_df}{igraph}(x, include_dyad = TRUE, leave_raw = FALSE)

\method{edg_to_df}{network}(x, include_dyad = TRUE, leave_raw = FALSE)

\method{edg_to_df}{tbl_graph}(x, include_dyad = TRUE,
  leave_raw = FALSE)

\method{print}{edge_data_frame}(x, .nrow = 10L)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{include_dyad}{\code{logical} (default: \code{TRUE}) indicating whether to return the \code{.ego}
and \code{.alter} of each edge's dyad.}

\item{leave_raw}{\code{logical} (default: \code{FALSE}) indicating whether to return a bare
\code{data.frame} (\code{stringsAsFactors = FALSE}) instead of an \code{edge_data_frame}.}

\item{.nrow}{Maximum number of rows to \code{print()}.}
}
\value{
\code{edge_data_frame} \code{data.frame} with all edge attributes contained in \code{x}.
}
\description{
Extract all edge attributes as a data frame.
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

edg_to_df(sampson_monastery)

edg_to_df(ig)

edg_to_df(nw)

edg_to_df(tidy_g)

}
\seealso{
\code{\link[=edg_get_attr]{edg_get_attr()}}, \code{\link[=edg_attr_names]{edg_attr_names()}}, \code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}

\code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
