% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edge_get_attr_names}
\alias{edge_get_attr_names}
\alias{edge_get_attr_names.igraph}
\alias{edge_get_attr_names.network}
\title{Get the names of a network's edge attributes.}
\usage{
edge_get_attr_names(.net)

\method{edge_get_attr_names}{igraph}(.net)

\method{edge_get_attr_names}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{character} \code{vector}
}
\description{
Get the names of a network's edge attributes.
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
