% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_.R
\name{ei_index_global}
\alias{ei_index_global}
\alias{ei_index_global.igraph}
\alias{ei_index_global.network}
\alias{ei_index_global.tbl_graph}
\alias{ei_index_global.attr_el}
\alias{ei_index_global.attr_adj_mat}
\alias{ei_index_global.mixing_matrix}
\title{Global E-I Index}
\usage{
ei_index_global(x, vrt_attr, ...)

\method{ei_index_global}{igraph}(x, vrt_attr, drop_loops = FALSE)

\method{ei_index_global}{network}(x, vrt_attr, drop_loops = FALSE)

\method{ei_index_global}{tbl_graph}(x, vrt_attr, ...)

\method{ei_index_global}{attr_el}(attr_el)

\method{ei_index_global}{attr_adj_mat}(x)

\method{ei_index_global}{mixing_matrix}(mixing_matrix)
}
\arguments{
\item{x}{\code{igraph}, \code{network}, \code{tbl_graph}, \code{attr_el}, \code{attr_adj_mat}, or \code{mixing_matrix}}

\item{vrt_attr}{\code{character} representing the name of a categorical vertex attribute
present in \code{x}.}

\item{drop_loops}{\code{logical}. Whether loops should be removed prior to calculations. \cr
Default: \code{FALSE}}
}
\value{
A scalar \code{double}.
}
\description{
Calculate a network's global E-I Index from various representations.
}
\examples{
library(snatools)

# igraph ==============================================================================
data("UKfaculty", package = "igraphdata")

ei_index_global(UKfaculty, vrt_attr = "Group")

# network =============================================================================
data("samplk", package = "ergm")

ei_index(samplk1, vrt_attr = "group")

# attr_adj_mat ========================================================================
(attr_adj_mat <- rep_attr_adj_mat(samplk1, vrt_attr = "group"))
ei_index_global(attr_adj_mat)

# attr_el =============================================================================
(attr_el <- rep_attr_el(attr_adj_mat)) \%>\% head()
ei_index_global(attr_el)

# mixing_matrix =======================================================================
(mix_mat <- rep_mixing_matrix(samplk1, vrt_attr = "group"))
ei_index_global(mix_mat)

# tbl_graph ===========================================================================
(tbl_g <- tidygraph::as_tbl_graph(UKfaculty))
ei_index_global(tbl_g, "Group")

}
\seealso{
\code{\link[=ei_index]{ei_index()}}. \code{\link[=rep_attr_el]{rep_attr_el()}}, \code{\link[=rep_attr_adj_mat]{rep_attr_adj_mat()}}, \code{\link[=rep_mixing_matrix]{rep_mixing_matrix()}}
}
