% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_.R
\name{ei_index_global_permute}
\alias{ei_index_global_permute}
\title{E-I Index Permutation Test}
\usage{
ei_index_global_permute(x, vrt_attr, iterations = 1000L,
  diagonal = FALSE)
}
\description{
E-I Index Permutation Test
}
\examples{
library(snatools)

data("samplk", package = "ergm")

res1 <- ei_index_global_permute(samplk1, "cloisterville")

res1

autoplot(res1)

res2 <- ei_index_global_permute(samplk1, "group")

res2

autoplot(res2)

}
