% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-structures.R
\name{example_adj_mat}
\alias{example_adj_mat}
\title{Build an Example Adjacency Matrix}
\usage{
example_adj_mat(directed = FALSE, loops = FALSE)
}
\arguments{
\item{directed}{\code{logical(1L)}, default: \code{FALSE}}

\item{loops}{\code{logical(1L)}, default: \code{FALSE}}
}
\description{
Build an Example Adjacency Matrix
}
\examples{
example_adj_mat()

example_adj_mat(directed = TRUE)

example_adj_mat(loops = TRUE)

example_adj_mat(directed = TRUE, loops = TRUE)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
