% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-structures.R
\name{example_el}
\alias{example_el}
\title{Build an Example Edge List}
\usage{
example_el(directed = FALSE, loops = FALSE)
}
\arguments{
\item{directed}{\code{logical(1L)}, default: \code{FALSE}}

\item{loops}{\code{logical(1L)}, default: \code{FALSE}}
}
\description{
Build an Example Edge List
}
\examples{
example_el()

example_el(directed = TRUE)

example_el(loops = TRUE)

example_el(directed = TRUE, loops = TRUE)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
