% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-structures.R
\name{example_igraph}
\alias{example_igraph}
\title{Build an Example \code{<igraph>}}
\usage{
example_igraph(directed = FALSE, loops = FALSE)
}
\arguments{
\item{directed}{\code{logical(1L)}, default: \code{FALSE}}

\item{loops}{\code{logical(1L)}, default: \code{FALSE}}
}
\description{
Build an Example \code{<igraph>}
}
\examples{
example_igraph()

example_igraph(directed = TRUE)

example_igraph(loops = TRUE)

example_igraph(directed = TRUE, loops = TRUE)

}
