% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-structures.R
\name{example_network}
\alias{example_network}
\title{Build an Example \code{<network>}}
\usage{
example_network(directed = FALSE, loops = FALSE,
  initial_rep = c("adj_mat", "el"))
}
\arguments{
\item{directed}{\code{logical(1L)}, default: \code{FALSE}}

\item{loops}{\code{logical(1L)}, default: \code{FALSE}}
}
\description{
Build an Example \code{<network>}
}
\examples{
example_network()

example_network(directed = TRUE)

example_network(loops = TRUE)

example_network(directed = TRUE, loops = TRUE)

}
