% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==\%}
\alias{\%==\%}
\title{Confirm \code{bridge_net}s, \code{igraph}s, and \code{network}s are \code{identical()}.}
\usage{
lhs \%==\% rhs
}
\arguments{
\item{lhs}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{rhs}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{logical} indicating whether \code{lhs} and \code{rhs} are \code{identical()}, excluding
non-structural network attributes.
}
\description{
Confirm \code{bridge_net}s, \code{igraph}s, and \code{network}s are \code{identical()}.
}
