% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jemmah_islamiyah.R
\docType{data}
\name{jemmah_islamiyah}
\alias{jemmah_islamiyah}
\title{Koschade's Jemaah Islamiyah (2002 Bali Bombings)}
\format{\itemize{
\item Object Class: \code{igraph}
\itemize{
\item directed: \code{FALSE}
\item bipartite: \code{FALSE}
\item multiplex: \code{FALSE}
\item vertices: \code{17}
\item edges: \code{63}
\item loops: \code{FALSE}
}
\item Vertex Attributes:
\itemize{
\item \code{name}
\itemize{
\item \code{character} representing each persons's name.
}
}
\item Edge Attributes:
\itemize{
\item \code{weight}
\itemize{
\item \code{numeric} describing the strength of relations between individuals, ranging from
\code{1} to \code{5}.
\itemize{
\item \code{1} refers to a weak relationship, such as a single text message or financial
transaction.
\item \code{5} refers to a strong relationship, such as individuals who resided together
or individuals who had numerous weak contacts over the period in question.
}
}
}
}}
\source{
\url{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/jemaahislamiyahkoschade}
}
\usage{
jemmah_islamiyah
}
\description{
\code{jemmah_islamiyah} is an adaptation of the \href{https://sites.google.com/site/ucinetsoftware/}{UCINET}
dataset "Jemaah Islamiyah Koschade". \cr
}
\details{
From the \href{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/jemaahislamiyahkoschade}{UCINET documentation}: \cr

Jemaah Islamiyah cell that was responsible for the Bali bombings
in 2002 – i.e. should tally with the other Bali bombing dataset from JJATT. The
recording of the interaction of the cell began following the  meeting in the Hotel
Harem in Denpasar on October 6, when the group was considered to go ‘operationally
covert’, and concluded when the majority of the group had left Bali before the
implementation of the operation on October 11, 2002.
}
\examples{
library(snatools)

jemmah_islamiyah

as_network(jemmah_islamiyah)

}
\references{
Koschade, Stuart (2006) A Social Network Analysis of Jemaah Islamiyah: The
Applications to Counter-Terrorism and Intelligence. Studies in Conflict and Terrorism
Vol. 29(6):pp. 559-575

International Crisis Group. 2003 ‘Jemaah Islamiyah in South East Asia:
Damaged but Still Dangerous’
}
\keyword{datasets}
