% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_is_.R
\name{net_is_bipartite}
\alias{net_is_bipartite}
\alias{net_is_bipartite.igraph}
\alias{net_is_bipartite.network}
\alias{net_is_directed}
\alias{net_is_directed.igraph}
\alias{net_is_directed.network}
\title{Metadata Queries}
\usage{
net_is_bipartite(x)

\method{net_is_bipartite}{igraph}(x)

\method{net_is_bipartite}{network}(x)

net_is_directed(x)

\method{net_is_directed}{igraph}(x)

\method{net_is_directed}{network}(x)
}
\arguments{
\item{x}{A graph object.}
}
\description{
Metadata Queries
}
\details{
\code{net_is_bipartite()} checks whether \code{x} is \emph{labeled} as bipartite and returns a \code{logical} scalar.

\code{net_is_directed()} checks whether \code{x} is \emph{labeled} as directed and returns a \code{logical} scalar.
}
\examples{
library(snatools)

data("karate", package = "igraphdata")
(ig <- karate)

data("emon", package = "network")
(nw <- emon$Cheyenne)

net_is_bipartite(ig)
net_is_bipartite(nw)

net_is_directed(ig)
net_is_directed(nw)
}
