% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_.R
\name{mix_ei_index}
\alias{mix_ei_index}
\alias{mix_ei_index.igraph}
\alias{mix_ei_index.network}
\alias{mix_ei_index.attr_el}
\title{E-I Index}
\usage{
mix_ei_index(x, vrt_attr, drop_loops = FALSE)

\method{mix_ei_index}{igraph}(x, vrt_attr, drop_loops = FALSE)

\method{mix_ei_index}{network}(x, vrt_attr, drop_loops = FALSE)

\method{mix_ei_index}{attr_el}(attr_el)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{vrt_attr}{\code{character} representing the name of a categorical vertex attribute present in \code{x}.}
}
\value{
\code{numeric} of length 1,  with \code{-1} suggesting homophilous tendencies and \code{+1} suggesting heterophilous tendencies.
}
\description{
The global ratio of external to internal ties.
}
\details{
\deqn{E-I~Index = \frac{EL-IL}{EL+IL}}
\itemize{
\item \eqn{EL} = the number of external ties.
\item \eqn{IL} = the number of inteneral ties.
}
}
\examples{
library(snatools)

# igraph ==========================================================================
data("UKfaculty", package = "igraphdata")

UKfaculty \%>\% 
  mix_ei_index(vrt_attr = "Group")

# network =========================================================================
data("sampson", package = "ergm")

samplike \%>\% 
  mix_ei_index("group")

}
\references{
Krackhardt, David, and Robert N. Stern. "Informal Networks and
Organizational Crises: An Experimental Simulation." Social Psychology Quarterly 51, no.
2 (1988): 123-40. \url{http://www.jstor.org/stable/2786835}.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
