% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_.R
\name{mix_mixing_matrix}
\alias{mix_mixing_matrix}
\alias{mix_mixing_matrix.igraph}
\alias{mix_mixing_matrix.network}
\title{Build a network/graph's mixing matrix.}
\usage{
mix_mixing_matrix(x, vrt_attr, drop_loops = FALSE)

\method{mix_mixing_matrix}{igraph}(x, vrt_attr, drop_loops = FALSE)

\method{mix_mixing_matrix}{network}(x, vrt_attr, drop_loops = FALSE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{vrt_attr}{\code{character} representing a vertex attribute.}
}
\description{
Build a network/graph's mixing matrix.
}
