% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monastery}
\alias{monastery}
\title{Sampson's Monastery}
\format{\code{list} of \code{igraph} objects}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm}
}
\usage{
monastery
}
\description{
Sampson's Monastery
}
\examples{
par(mfrow = c(2, 5))

snatools::monastery \%>\% 
  purrr::iwalk(~ plot(.x, main = .y, asp = 0, edge.arrow.size = 0.25,
               layout = igraph::layout_with_fr))

}
\references{
\itemize{
\item Breiger R., Boorman S. and Arabie P. (1975). An algorithm for clustering relational
data with applications to social network analysis and comparison with
multidimensional scaling. Journal of Mathematical Psychology, 12, 328-383.
\item Sampson, S. (1969). Crisis in a cloister. Unpublished doctoral dissertation, Cornell
University.
}
}
\keyword{datasets}
