% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_as_undirected}
\alias{net_as_undirected}
\alias{net_as_undirected.network}
\alias{net_as_undirected.igraph}
\title{Convert a directed graph object to undirected.}
\usage{
net_as_undirected(x, ...)

\method{net_as_undirected}{network}(x, ...)

\method{net_as_undirected}{igraph}(x, ...)
}
\description{
Convert a directed graph object to undirected.
}
\examples{
library(snatools)

ig <- c(0, 1, 0, 1,
        1, 0, 0, 0,
        0, 1, 0, 1,
        1, 0, 1, 0) \%>\%
  matrix(nrow = 4, byrow = TRUE) \%>\%
  igraph::graph_from_adjacency_matrix()

nw <- ig \%>\% 
  as_network()

nw

nw \%>\% 
  net_as_undirected()

ig

ig \%>\% 
  net_as_undirected()

}
\seealso{
\code{\link[igraph:as.undirected]{igraph::as.undirected()}}
}
