% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_.R
\name{net_attr_names}
\alias{net_attr_names}
\alias{net_attr_names.igraph}
\alias{net_attr_names.network}
\title{List a network/graph's attribute names.}
\usage{
net_attr_names(x, exclude_metadata = TRUE)

\method{net_attr_names}{igraph}(x, exclude_metadata = NULL)

\method{net_attr_names}{network}(x, exclude_metadata = TRUE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{exclude_metadata}{\code{logical} objects include what can be described as
"metadata" in their network/graph-level attributes.
\itemize{
\item Default: \code{TRUE}
These metadata include:
\item \code{"n"}: The number of vertices in the \code{network} object.
\item \code{"directed"}: Whether the \code{network} object is directed.
\item \code{"hyper"}: Whether the \code{network} object is a hypergraph.
\item \code{"multiple"}: Whether the \code{network} object is multiplex.
\item \code{"bipartite"}: Whether the \code{network} object is bipartite.
\item \code{"mnext"}: The index of the next edge to be added to the \code{network} object.
\code{exclude_metadata} determines whether or not to include these in the returned names.
\item If \code{exclude_metadata} is provided when \code{x} is an \code{igraph} object, it is ignored with
a \code{warning}.
}}
}
\value{
\code{character} \code{vector}of \code{x}'s network/graph attribute names.
}
\description{
List a network/graph's attribute names.
}
\seealso{
\code{\link[igraph:edge_attr_names]{igraph::edge_attr_names()}}, \code{\link[network:list.edge.attributes]{network::list.edge.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
