% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-attributes.R
\name{net_attrs_to_list}
\alias{net_attrs_to_list}
\alias{net_attrs_to_list.bridge_net}
\alias{net_attrs_to_list.igraph}
\alias{net_attrs_to_list.network}
\alias{net_attrs_to_list.tbl_graph}
\title{Extract all non-structural network attributes as a named \code{list}.}
\usage{
net_attrs_to_list(x)

\method{net_attrs_to_list}{bridge_net}(x)

\method{net_attrs_to_list}{igraph}(x)

\method{net_attrs_to_list}{network}(x)

\method{net_attrs_to_list}{tbl_graph}(x)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
A named \code{list()} of all of \code{x}'s network attributes.
}
\description{
Extract all non-structural network attributes as a named \code{list}.
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

net_attrs_to_list(sampson_monastery)

net_attrs_to_list(ig)

net_attrs_to_list(nw)

net_attrs_to_list(tidy_g)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
