% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_count_actors}
\alias{net_count_actors}
\alias{net_count_actors.igraph}
\alias{net_count_actors.network}
\alias{net_count_actors.bridge_net}
\alias{net_count_actors.tbl_graph}
\alias{net_count_actors.edgelist}
\alias{net_count_actors.adj_matrix}
\title{Count the number of vertices belonging to the actor mode of a bipartite graph object.}
\usage{
net_count_actors(x)

\method{net_count_actors}{igraph}(x)

\method{net_count_actors}{network}(x)

\method{net_count_actors}{bridge_net}(x)

\method{net_count_actors}{tbl_graph}(x)

\method{net_count_actors}{edgelist}(x)

\method{net_count_actors}{adj_matrix}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{integer} corresponding to the number \code{x}'s actors.
}
\description{
Count the number of vertices belonging to the actor mode of a bipartite graph object.
}
\examples{
library(snatools)

ig <- igraph::bipartite.random.game(7, 3, type = "gnp", p = 0.4)

nw <- ig \%>\% 
  as_network()
  
bridge_g <- ig \%>\% 
  as_bridge_net()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()
  
el <- ig \%>\% 
  rep_as_edgelist(use_names = FALSE)

net_count_actors(ig)

net_count_actors(nw)

net_count_actors(bridge_g)

net_count_actors(tidy_g)

net_count_actors(el)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
