% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_count_edges}
\alias{net_count_edges}
\alias{net_count_edges.sna_net}
\alias{net_count_edges.igraph}
\alias{net_count_edges.network}
\alias{net_count_edges.tbl_graph}
\alias{net_count_edges.edgelist}
\title{Count the number of edges present in a graph object.}
\usage{
net_count_edges(x)

\method{net_count_edges}{sna_net}(x)

\method{net_count_edges}{igraph}(x)

\method{net_count_edges}{network}(x)

\method{net_count_edges}{tbl_graph}(x)

\method{net_count_edges}{edgelist}(x)
}
\arguments{
\item{x}{An \code{sna_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{integer} indicating the number of edges in \code{x}.
}
\description{
Count the number of edges present in a graph object.
}
\examples{
library(snatools)

ig <- igraph::random.graph.game(10, 0.25, directed = TRUE)

nw <- ig \%>\% 
  as_network()
  
sna_g <- ig \%>\% 
  as_sna_net()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()
  
el <- ig \%>\% 
  rep_as_edgelist(use_names = FALSE)

net_count_edges(sna_g)

net_count_edges(ig)

net_count_edges(nw)

net_count_edges(tidy_g)

net_count_edges(el)

}
\seealso{
\code{\link[igraph:ecount]{igraph::ecount()}}

\code{\link[network:network.edgecount]{network::network.edgecount()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
