% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_count_vertices}
\alias{net_count_vertices}
\alias{net_count_vertices.bridge_net}
\alias{net_count_vertices.igraph}
\alias{net_count_vertices.network}
\alias{net_count_vertices.tbl_graph}
\alias{net_count_vertices.edgelist}
\alias{net_count_vertices.adj_matrix}
\title{Count the number of vertices present in a graph object.}
\usage{
net_count_vertices(x)

\method{net_count_vertices}{bridge_net}(x)

\method{net_count_vertices}{igraph}(x)

\method{net_count_vertices}{network}(x)

\method{net_count_vertices}{tbl_graph}(x)

\method{net_count_vertices}{edgelist}(x)

\method{net_count_vertices}{adj_matrix}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{integer} indicating the number of vertices in \code{x}.
}
\description{
Count the number of vertices present in a graph object.
}
\examples{
library(snatools)

ig <- igraph::random.graph.game(10, 0.25, directed = TRUE)

nw <- ig \%>\% 
  as_network()
  
bridge_g <- ig \%>\% 
  as_bridge_net()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()
  
el <- ig \%>\% 
  rep_as_edgelist(use_names = FALSE)
  
adj_matrix <- ig \%>\% 
 rep_as_adj_matrix()

net_count_vertices(bridge_g)

net_count_vertices(ig)

net_count_vertices(nw)

net_count_vertices(tidy_g)

net_count_vertices(el)

net_count_vertices(el)

}
\seealso{
\code{\link[igraph:vcount]{igraph::vcount()}}

\code{\link[network:network.size]{network::network.size()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
