% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-attributes.R
\name{net_get_attr}
\alias{net_get_attr}
\alias{net_get_attr.bridge_net}
\alias{net_get_attr.igraph}
\alias{net_get_attr.network}
\alias{net_get_attr.tbl_graph}
\title{Extract a specific, non-strucural network attribute.}
\usage{
net_get_attr(x, net_attr)

\method{net_get_attr}{bridge_net}(x, net_attr)

\method{net_get_attr}{igraph}(x, net_attr)

\method{net_get_attr}{network}(x, net_attr)

\method{net_get_attr}{tbl_graph}(x, net_attr)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{net_attr}{\code{character} specifying the target network attribute.}
}
\description{
Extract a specific, non-strucural network attribute.
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

sampson_monastery \%>\% 
  net_get_attr("network_name")

ig \%>\% 
  net_get_attr("author")

tidy_g \%>\% 
  net_get_attr("author")

}
\seealso{
\code{\link[igraph:graph_attr]{igraph::graph_attr()}}

\code{\link[network:get.network.attribute]{network::get.network.attribute()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
