% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_get.R
\name{net_get_attrs}
\alias{net_get_attrs}
\alias{net_get_attrs.igraph}
\alias{net_get_attrs.network}
\title{Extract graph attributes.}
\usage{
net_get_attrs(x)

\method{net_get_attrs}{igraph}(ig)

\method{net_get_attrs}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s graph attributes.
}
\description{
Extract graph attributes.
}
