% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_has_isolates}
\alias{net_has_isolates}
\alias{net_has_isolates.bridge_net}
\alias{net_has_isolates.igraph}
\alias{net_has_isolates.network}
\alias{net_has_isolates.tbl_graph}
\alias{net_has_isolates.edgelist}
\title{Does a graph object contain isolates?}
\usage{
net_has_isolates(x)

\method{net_has_isolates}{bridge_net}(x)

\method{net_has_isolates}{igraph}(x)

\method{net_has_isolates}{network}(x)

\method{net_has_isolates}{tbl_graph}(x)

\method{net_has_isolates}{edgelist}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{logical} indicating whether \code{x} contains isolated vertices.
}
\description{
Test whether any vertices have 0 edges.
}
\examples{
library(snatools)

ig <- igraph::make_empty_graph(n = 5) \%>\% 
  igraph::add_edges(c(1, 2, 2, 3, 4, 1, 2, 4))

nw <- ig \%>\% 
  as_network()
  
bridge_g <- ig \%>\% 
  as_bridge_net()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()
  
el <- ig \%>\% 
  rep_as_edgelist(use_names = FALSE)

net_has_isolates(bridge_g)

net_has_isolates(ig)

net_has_isolates(nw)

net_has_isolates(tidy_g)

net_has_isolates(el)

}
\seealso{
\code{\link[tidygraph:node_is_isolated]{tidygraph::node_is_isolated()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
