% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_is_bipartite}
\alias{net_is_bipartite}
\alias{net_is_bipartite.igraph}
\alias{net_is_bipartite.network}
\alias{net_is_bipartite.tbl_graph}
\title{Is a Graph Bipartite?}
\usage{
net_is_bipartite(x)

\method{net_is_bipartite}{igraph}(x)

\method{net_is_bipartite}{network}(x)

\method{net_is_bipartite}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{logical} indicating whether \code{x} is a bipartite graph.
}
\description{
Is a Graph Bipartite?
}
\seealso{
\code{\link[igraph:is_bipartite]{igraph::is_bipartite()}}

\code{\link[network:is.bipartite]{network::is.bipartite()}}

\code{\link[tidygraph:graph_is_bipartite]{tidygraph::graph_is_bipartite()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
