% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_is_directed}
\alias{net_is_directed}
\alias{net_is_directed.igraph}
\alias{net_is_directed.network}
\alias{net_is_directed.tbl_graph}
\title{Is a Graph Directed?}
\usage{
net_is_directed(x)

\method{net_is_directed}{igraph}(x)

\method{net_is_directed}{network}(x)

\method{net_is_directed}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{logical} indicating whether \code{x} is a directed graph.
}
\description{
Is a Graph Directed?
}
\seealso{
\code{\link[igraph:is_directed]{igraph::is_directed()}}

\code{\link[network:is.directed]{network::is.directed()}}

\code{\link[tidygraph:graph_is_directed]{tidygraph::graph_is_directed()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
