% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-node.R
\name{node_attr_names}
\alias{node_attr_names}
\alias{node_attr_names.igraph}
\alias{node_attr_names.network}
\title{Get the Names of Node Attributes}
\usage{
node_attr_names(x, ...)

\method{node_attr_names}{igraph}(x, ...)

\method{node_attr_names}{network}(x, ignore_na = TRUE, ...)
}
\arguments{
\item{x}{\code{<igraph>} or \code{<network>}}

\item{...}{Arguments passed to or from other methods.}

\item{ignore_na}{\code{logical(1L)}, default: \code{TRUE}. Whether to ignore the \code{"na"}
attribute that \code{{network}} automatically creates.}
}
\value{
\code{character()}
}
\description{
Get the Names of Node Attributes
}
\examples{
ig <- example_igraph()
node_attr_names(ig)

nw <- example_network()
node_attr_names(nw)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
