% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-node.R
\name{node_count}
\alias{node_count}
\alias{node_count.igraph}
\alias{node_count.network}
\title{How Many Nodes Are There?}
\usage{
node_count(x)

\method{node_count}{igraph}(x)

\method{node_count}{network}(x)
}
\arguments{
\item{x}{\code{<igraph>} or \code{<network>}}
}
\value{
\code{<integer>}
}
\description{
How Many Nodes Are There?
}
\examples{
ig <- example_igraph()
node_count(ig)

nw <- example_network()
node_count(nw)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
