% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-node.R
\name{node_get_attr}
\alias{node_get_attr}
\alias{node_get_attr.igraph}
\alias{node_get_attr.network}
\title{Get Node Attributes}
\usage{
node_get_attr(x, attr_name, which_nodes = seq_nodes(x), ...)

\method{node_get_attr}{igraph}(x, attr_name, which_nodes = seq_nodes(x),
  ...)

\method{node_get_attr}{network}(x, attr_name, which_nodes = seq_nodes(x),
  auto_unlist = TRUE, ...)
}
\arguments{
\item{x}{\code{<igraph>} or \code{<network>}}

\item{attr_name}{\code{<character>} (scalar)}

\item{which_nodes}{\code{<numeric>} or \code{<logical>}}

\item{...}{Arguments passed to or from other methods.}

\item{auto_unlist}{If possible, safely flatten the result to an atomic vector.}
}
\value{
\code{atomic} \code{vector} or \code{<list>}
}
\description{
Get Node Attributes
}
\examples{
ig <- example_igraph()
node_count(ig)

nw <- example_network()
node_count(nw)

}
