% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-node.R
\name{node_get_names}
\alias{node_get_names}
\alias{node_get_names.igraph}
\alias{node_get_names.network}
\title{Get Node Names}
\usage{
node_get_names(x, ...)

\method{node_get_names}{igraph}(x, ...)

\method{node_get_names}{network}(x, ...)
}
\arguments{
\item{x}{\code{<igraph>} or \code{<network>}}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Get Node Names
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
