% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_matrix.R
\name{permute_matrix}
\alias{permute_matrix}
\alias{permute_matrix.igraph}
\alias{permute_matrix.network}
\alias{permute_matrix.matrix}
\title{Permute matrix columns and rows.}
\usage{
permute_matrix(x, ...)

\method{permute_matrix}{igraph}(x, matrix_fun = rep_adjacency_matrix,
  ...)

\method{permute_matrix}{network}(x, matrix_fun = rep_adjacency_matrix,
  ...)

\method{permute_matrix}{matrix}(x, out_class = class(x))
}
\description{
Permute matrix columns and rows.
}
