% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snatools-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Ceci n'est pas une pipe.}
\description{
{\code{snatools}} encourages use of \code{\%>\%} from {\code{magrittr}}.
}
\seealso{
\code{\link[magrittr:\%>\%]{magrittr::\%>\%()}}
}
