% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_triad_census}
\alias{plot_triad_census}
\alias{plot_triad_census.igraph}
\title{Quickly plot a graph's triad census.}
\usage{
plot_triad_census(x, coord_flip = TRUE, pal = rainbow)

\method{plot_triad_census}{igraph}(x, coord_flip = TRUE, pal = rainbow)
}
\description{
Quickly plot a graph's triad census.
}
\examples{
library(snatools)

data("UKfaculty", package = "igraphdata")

UKfaculty \%>\% 
  plot_triad_census(coord_flip = FALSE)

}
\seealso{
\code{\link[igraph:triad_census]{igraph::triad_census()}}, \code{\link[sna:triad.census]{sna::triad.census()}}
}
