% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qap.R
\name{qap_cor}
\alias{qap_cor}
\alias{qap_cor.default}
\alias{qap_cor.adj_mat_pair}
\title{Quadratic Assignment Procedure (Correlation)}
\usage{
qap_cor(g1, g2, iterations = 1000L, diagonal = FALSE)

\method{qap_cor}{default}(x, y, iterations = NULL, diagonal = FALSE)

\method{qap_cor}{adj_mat_pair}(adj_mat_pair, iterations,
  diagonal = FALSE)
}
\description{
Quadratic Assignment Procedure (Correlation)
}
\examples{
library(snatools)

data("florentine", package = "ergm")

res <- qap_cor(flomarriage, flobusiness)

res

autoplot(res)

}
