% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ucinet.R
\name{read_ucinet}
\alias{read_ucinet}
\title{Read UCINET's .dl files}
\usage{
read_ucinet(path, directed = NULL)
}
\arguments{
\item{path}{Path to file or URL.}
}
\value{
A \code{ucinet} object.
}
\description{
Read UCINET's .dl files
}
