% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\title{Visualization}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

