% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_adjacency_matrix}
\alias{rep_adjacency_matrix}
\alias{rep_adjacency_matrix.igraph}
\alias{rep_adjacency_matrix.network}
\title{Build a network's adjacency matrix representation}
\usage{
rep_adjacency_matrix(x, ...)

\method{rep_adjacency_matrix}{igraph}(x, edg_attr = NULL,
  sparse = FALSE, ...)

\method{rep_adjacency_matrix}{network}(x, edg_attr = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{
\code{matrix} or \code{dgCMatrix}
}
\description{
Build a network's adjacency matrix representation
}
\examples{
library(snatools)

data("sampson", package = "ergm")

rep_adjacency_matrix(samplike)

}
