% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_list.R
\name{rep_as_adj_list}
\alias{rep_as_adj_list}
\title{Construct a graph object's adjacency list representation.}
\usage{
rep_as_adj_list(x, use_names = TRUE, vrt_attr = NULL)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{use_names}{\code{logical} (default: \code{TRUE}) indicating whether to use vertex names as
the values filling the returned object. If \code{FALSE}, or a relevant attribute does not
exist, the returned object uses vertex indices.}

\item{vrt_attr}{\code{character} specifying the target vertex attribute.}
}
\value{
\code{list}, where \code{names} and values correspond to vertex names, indices,
or attributes, depending on arguments provided.
}
\description{
Construct a graph object's adjacency list representation.
}
\examples{
library(snatools)

jemmah_islamiyah \%>\% 
  rep_as_adj_list()
  
jemmah_islamiyah \%>\% 
  rep_as_adj_list(use_names = FALSE)

}
\seealso{
\code{\link[igraph:as_adj_list]{igraph::as_adj_list()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
