% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency-matrix.R
\name{rep_as_adj_matrix}
\alias{rep_as_adj_matrix}
\alias{print.adj_matrix}
\title{Construct a graph object's adjacency matrix representation.}
\usage{
rep_as_adj_matrix(x, use_names = TRUE, vrt_attr = NULL,
  edg_attr = NULL, leave_raw = FALSE)

\method{print}{adj_matrix}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, or\code{tbl_graph}, or \code{edgelist} object.}

\item{use_names}{\code{logical} (default: \code{TRUE}) indicating whether to use vertex names for
row and column names.}

\item{vrt_attr}{\code{character} (default: \code{NULL}) indicating which vertex attribute to use
in for row and column names. If provided, \code{vrt_attr} overrides \code{use_names}.`}

\item{sparse}{\code{logical} (default: \code{TRUE}) indicating whether to use
\code{Matrix::sparseMatrix()} in constructing the adjacency matrix.}
}
\value{
An \code{adj_matrix} object.
}
\description{
Construct a graph object's adjacency matrix representation.
}
\examples{
library(snatools)

data(sampson_monastery)
  
# sampson_monastery \%>\% 
# edg_filter(time == 1) \%>\% 
#   rep_as_adj_matrix()

sampson_monastery \%>\% 
  edg_filter(time == 1) \%>\% 
  as_igraph() \%>\% 
  rep_as_adj_matrix(vrt_attr = "status")
  
sampson_monastery \%>\% 
  edg_filter(time == 2) \%>\% 
  as_network() \%>\% 
  rep_as_adj_matrix(edg_attr = "weight")
  
sampson_monastery \%>\% 
  edg_filter(time == 3, relation == "blame") \%>\% 
  tidygraph::as_tbl_graph() \%>\% 
  rep_as_adj_matrix(vrt_attr = "faction")

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
