% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_as.R
\name{rep_as_adjacency_matrix}
\alias{rep_as_adjacency_matrix}
\alias{rep_as_adjacency_matrix.igraph}
\alias{rep_as_adjacency_matrix.network}
\title{Build a network's adjacency matrix representation}
\usage{
rep_as_adjacency_matrix(x, edg_attr = NULL, sparse = FALSE, ...)

\method{rep_as_adjacency_matrix}{igraph}(x, edg_attr = NULL,
  sparse = FALSE, ...)

\method{rep_as_adjacency_matrix}{network}(x, edg_attr = NULL,
  sparse = FALSE)
}
\arguments{
\item{...}{}
}
\value{
\code{matrix} or \code{dgCMatrix}
}
\description{
Build a network's adjacency matrix representation
}
