% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_as.R
\name{rep_as_edgelist}
\alias{rep_as_edgelist}
\alias{rep_as_edgelist.igraph}
\alias{rep_as_edgelist.network}
\title{Build a networks's edge list representation.}
\usage{
rep_as_edgelist(x, use_names = FALSE)

\method{rep_as_edgelist}{igraph}(ig, use_names = FALSE)

\method{rep_as_edgelist}{network}(nw, use_names = FALSE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{use_names}{\code{logical} Whether the returned edge list should use vertex names instead of their
indices. \cr
Default: \code{FALSE}}
}
\value{
\code{matrix} \cr
\itemize{
\item \code{mode()}:
\itemize{
\item \code{numeric} if \code{use_names} is \code{FALSE}.
\item \code{character} if \code{use_names} is \code{TRUE}.
}
\item \code{dim()}
\itemize{
\item \code{igraph::ecount(x)} by 2 if \code{x} is an \code{igraph} object.
\item \code{network::network.edgecount(x)} by 2 if \code{x} is a \code{network} object.
}
}
}
\description{
\code{rep_as_edgelist()} creates edge lists in a \emph{consistent} format across classes.
}
\details{
\itemize{
\item If \code{x} is \emph{directed}, columns are named \code{"from"} and \code{"to"} to indicate the
tie's direction.
\item If \code{x} is \emph{undirected}, columns are named \code{"vert1"} and \code{"vert2"}.
}
}
\examples{
library(snatools)

# igraph ==============================================================================
data("karate", package = "igraphdata") # loads "Zachary's Karate Club" as `karate`
karate

karate \%>\% 
  rep_as_edgelist() \%>\% 
  head()
  
karate \%>\% 
  rep_as_edgelist(use_names = TRUE) \%>\% 
  head()

# network =============================================================================
data("sampson", package = "ergm") # loads "Sampson's Monks" as `samplike`
samplike

samplike \%>\% 
  rep_as_edgelist() \%>\% 
  head()
  
samplike \%>\% 
  rep_as_edgelist(use_names = TRUE) \%>\% 
  head()

}
\seealso{
\code{\link[igraph:as_edgelist]{igraph::as_edgelist()}}, \code{\link[network:as.edgelist]{network::as.edgelist()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
