% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_matrix.R
\name{rep_as_mixing_matrix}
\alias{rep_as_mixing_matrix}
\title{Construct a graph object's mixing matrix.}
\usage{
rep_as_mixing_matrix(x, vrt_attr)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{vrt_attr}{\code{character} specifying the target vertex attribute.}
}
\description{
Using a categorical vertex attribute, construct a matrix depicting network mixing
patterns.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
