% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_attr_adj_mat}
\alias{rep_attr_adj_mat}
\alias{rep_attr_adj_mat.igraph}
\alias{rep_attr_adj_mat.network}
\title{Build a graphs's adjacency matrix, using an attribute category in place of vertex names.}
\usage{
rep_attr_adj_mat(x, vrt_attr)

\method{rep_attr_adj_mat}{igraph}(x, vrt_attr)

\method{rep_attr_adj_mat}{network}(x, vrt_attr)
}
\description{
Build a graphs's adjacency matrix, using an attribute category in place of vertex names.
}
