% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_biadjacency_matrix}
\alias{rep_biadjacency_matrix}
\alias{rep_biadjacency_matrix.igraph}
\alias{rep_biadjacency_matrix.network}
\title{rep_biadjacency_matrix}
\usage{
rep_biadjacency_matrix(x, ...)

\method{rep_biadjacency_matrix}{igraph}(x, ...)

\method{rep_biadjacency_matrix}{network}(x, ...)
}
\description{
rep_biadjacency_matrix
}
