% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_mixing_matrix}
\alias{rep_mixing_matrix}
\alias{rep_mixing_matrix.igraph}
\alias{rep_mixing_matrix.network}
\title{Build a network/graph's mixing matrix.}
\usage{
rep_mixing_matrix(x, vrt_attr, drop_loops = FALSE)

\method{rep_mixing_matrix}{igraph}(x, vrt_attr, drop_loops = FALSE)

\method{rep_mixing_matrix}{network}(x, vrt_attr, drop_loops = FALSE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{vrt_attr}{\code{character} representing a vertex attribute.}
}
\description{
Build a network/graph's mixing matrix.
}
