% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_two_mode_matrix}
\alias{rep_two_mode_matrix}
\alias{rep_two_mode_matrix.igraph}
\alias{rep_two_mode_matrix.network}
\title{rep_two_mode_matrix}
\usage{
rep_two_mode_matrix(x, edg_attr = NULL, sparse = FALSE, ...)

\method{rep_two_mode_matrix}{igraph}(x, edg_attr = NULL,
  sparse = FALSE, ...)

\method{rep_two_mode_matrix}{network}(x, edg_attr = NULL,
  sparse = FALSE, ...)
}
\description{
rep_two_mode_matrix
}
