% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-node.R
\name{seq_nodes}
\alias{seq_nodes}
\title{Get a Sequence of Nodes for Indexing}
\usage{
seq_nodes(x)
}
\arguments{
\item{x}{\code{<igraph>} or \code{<network>}}
}
\value{
\code{<integer>}
}
\description{
Get a Sequence of Nodes for Indexing
}
\examples{
ig <- example_igraph()
seq_nodes(ig)

nw <- example_network()
seq_nodes(nw)

}
\seealso{
\code{\link[igraph:V]{igraph::V()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
