% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_whole_number_storage}
\alias{set_whole_number_storage}
\title{#' @importFrom rlang is_named
sort_list_by_name <- function(x) {
if(is_named(x)) {
return(x\link{order(names(x))})
}
x
}}
\usage{
set_whole_number_storage(x, big_tol = .Machine$integer.max * 0.8,
  verbose = FALSE)
}
\description{
clear_metadata <- function(x) {
metadata <- c("n", "directed", "hyper", "loops", "multiple", "bipartite", "mnext")
out <- x\link{!names(x) \%in\% metadata}
sort_list_by_name(out)
}
}
