% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_as_igraph.R
\name{sna_as_igraph}
\alias{sna_as_igraph}
\alias{sna_as_igraph.network}
\title{Convert \code{network} objects to \code{igraph}.}
\usage{
sna_as_igraph(x)

\method{sna_as_igraph}{network}(nw)
}
\description{
Convert \code{network} objects to \code{igraph}.
}
