% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_as_network.R
\name{sna_as_network}
\alias{sna_as_network}
\alias{sna_as_network.igraph}
\title{Convert a graph object into a \code{network} object.}
\usage{
sna_as_network(x)

\method{sna_as_network}{igraph}(ig)
}
\arguments{
\item{x}{A graph object.\code{igraph} objects are supported.}
}
\value{
A \code{network} object.
}
\description{
Convert a graph object into a \code{network} object.
}
