% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_clean_graph.R
\name{sna_clean_graph}
\alias{sna_clean_graph}
\alias{sna_clean_graph.igraph}
\alias{sna_clean_graph.network}
\title{Clean a graph object's attributes in a standardized fashion.}
\usage{
sna_clean_graph(x)

\method{sna_clean_graph}{igraph}(ig)

\method{sna_clean_graph}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A cleaned version of \code{x}.
}
\description{
Clean a graph object's attributes in a standardized fashion.
}
\examples{
library(snatools)

data("Koenigsberg", package = "igraphdata")

Koenigsberg \%>\% sna_get_vert_attrs()

Koenigsberg \%>\% sna_clean_graph() \%>\% sna_get_vert_attrs()


}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
