% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_get_edge_attrs.R
\name{sna_get_edge_attrs}
\alias{sna_get_edge_attrs}
\alias{sna_get_edge_attrs.igraph}
\alias{sna_get_edge_attrs.network}
\title{Extract edge attributes from a graph.}
\usage{
sna_get_edge_attrs(x)

\method{sna_get_edge_attrs}{igraph}(ig)

\method{sna_get_edge_attrs}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s edge attributes.
}
\description{
Extract edge attributes from a graph.
}
