% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_get_graph_attrs.R
\name{sna_get_graph_attrs}
\alias{sna_get_graph_attrs}
\title{Extract graph-level attributes from a graph.}
\usage{
sna_get_graph_attrs(x)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s graph attributes.
}
\description{
Extract graph-level attributes from a graph.
}
