% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_get_vert_attrs.R
\name{sna_get_vert_attrs}
\alias{sna_get_vert_attrs}
\alias{sna_get_vert_attrs.igraph}
\alias{sna_get_vert_attrs.network}
\title{Extract vertex attributes from a graph.}
\usage{
sna_get_vert_attrs(x)

\method{sna_get_vert_attrs}{igraph}(ig)

\method{sna_get_vert_attrs}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s vertex attributes.
}
\description{
Extract vertex attributes from a graph.
}
