% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_net.R
\name{sna_net}
\alias{sna_net}
\alias{as_sna_net}
\alias{as_sna_net.sna_net}
\alias{as_sna_net.network}
\alias{as_sna_net.igraph}
\alias{as_sna_net.tbl_graph}
\title{The \code{sna_net} object effectively bridges the \code{network} objects used by the \code{statnet}
suite and \code{igraph} objects and their derivatives.}
\usage{
sna_net(edges = NULL, vertices = NULL, is_directed = FALSE)

as_sna_net(x, ...)

\method{as_sna_net}{sna_net}(x)

\method{as_sna_net}{network}(x, edg_attrs = TRUE, vrt_attrs = TRUE,
  net_attrs = TRUE)

\method{as_sna_net}{igraph}(x, edg_attrs = TRUE, vrt_attrs = TRUE,
  net_attrs = TRUE)

\method{as_sna_net}{tbl_graph}(x, edg_attrs = TRUE, vrt_attrs = TRUE,
  net_attrs = TRUE)
}
\arguments{
\item{x}{\code{igraph} (\code{\link[igraph:graph]{igraph::graph}}), \code{\link[network:network]{network::network}}, or \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}}.}

\item{...}{Named arguments passed on to other methods. Currently \code{edg_attrs},
\code{vrt_attrs}, and \code{net_attrs}.}

\item{edg_attrs}{\code{logical}, defaults \code{TRUE}. Whether to retain \code{x}'s edge attributes.}

\item{vrt_attrs}{\code{logical}, defaults \code{TRUE}. Whether to retain \code{x}'s vertex attributes.}

\item{net_attrs}{\code{logical}, defaults \code{TRUE}. Whether to retain \code{x}s network/graph-level
attributes.}
}
\description{
The \code{sna_net} object effectively bridges the \code{network} objects used by the \code{statnet}
suite and \code{igraph} objects and their derivatives.
}
\details{
\code{sna_net} objects contain the following elements: \cr
\itemize{
\item \code{graph_attributes}
\itemize{
\item \code{graph_attributes} describe external data fields used to annotate objects, such as
graph names or citation information.
}
\item \code{metadata}
\itemize{
\item \code{metadata} describe the graph itself and include:
\itemize{
\item \code{n_vertices}: \code{integer} counting the number of vertices in the graph.
\item \code{n_edges}: \code{integer} counting the number of edges in the graph.
\item \code{is_directed}: \code{logical} indicating whether edges are intended to be directed.
\item \code{is_bipartite}: \code{logical} indicating whether the graph is biparite
\itemize{
\item \code{n_actors}: \code{integer} counting the number of vertices belonging to a bipartite
graph's "actor" mode.
}
\item \code{any_multiplex}: \code{logical} indicating whether the graph includes any parallel
edges.
\item \code{any_loops}: \code{logical} indicating whether the graph contains loop edges.
\item \code{any_isolates}: \code{logical} indicating whether the graph contains any isolated
vertices.
}
}
\item \code{edges}
\itemize{
\item \code{edges} are contained in either a \code{matrix} or \code{data.frame} where the first column
(\code{.ego}) always contains an edge's source and the second column (\code{.alter}) always
contains an edge's target.
\itemize{
\item \code{.ego} and \code{.alter} are stored as type \code{integer}.
}
\item If the edge data of an \code{sna_net} object only include \code{.ego}s and \code{.alter}s, \code{edges}
are stored in a two-column \code{matrix}.
\item If edge attribute data are present, \code{edges} are stored in a \code{data.frame} where
each additional column contains a given edge attribute.
}
\item \code{vertices}
\itemize{
\item \code{vertices} are contained in a \code{data.frame} where the first column always
contains vertex \code{.name}s and subsequent columns contain any vertex attributes
present.
\itemize{
\item If an \code{sna_net} contains a bipartite graph, the second column will always
be \code{.actor}, a \code{logical} \code{vector} indicating whether each vertex is intended
to belong to the "actor" mode.
}
}
}
}
