% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_standardize_graph.R
\name{sna_standardize_graph}
\alias{sna_standardize_graph}
\title{Clean a graph object's attributes in a standardized fashion.}
\usage{
sna_standardize_graph(x)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A cleaned version of \code{x}.
}
\description{
Clean a graph object's attributes in a standardized fashion.
}
