% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southern_women}
\alias{southern_women}
\title{Davis' Southern Women}
\format{\code{igraph}}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm}
}
\usage{
southern_women
}
\description{
Davis' Southern Women
}
\examples{

}
\references{
\itemize{
\item Breiger R. (1974). The duality of persons and groups. Social Forces, 53, 181-190.
\item Davis, A et al. (1941). Deep South. Chicago: University of Chicago Press.
}
}
\keyword{datasets}
