% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%==\%.default}
\alias{\%==\%.igraph}
\title{Strict object comparison.}
\usage{
lhs \%==\% rhs

\method{\%==\%}{default}(lhs, rhs)

\method{\%==\%}{igraph}(lhs, rhs)
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand side}
}
\value{
\code{logical} Whether \code{lhs} and \code{rhs} are identical.
}
\description{
\itemize{
\item This infix operator
\itemize{
\item provides stricter comparisons than \code{==}.
\item prevents silent type conversions.
\item offers a method to easily compare \code{igraph} objects.
}
}
}
\examples{

TRUE == 1
TRUE \%==\% 1

NA == NA_character_
NA \%==\% NA_character_

NULL == NA_integer_
NULL \%==\% NA_integer_

NA == "NA"
NA \%==\% "NA"

igraph::graph("Zachary") \%==\% igraph::graph("Zachary")

}
\seealso{
\code{\link[base:identical]{base::identical()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
