% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictify.R
\name{strictify}
\alias{strictify}
\alias{strictify.igraph}
\alias{strictify.network}
\title{Clean a graph object's attributes in a standardized fashion.}
\usage{
strictify(x)

\method{strictify}{igraph}(ig)

\method{strictify}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A cleaned version of \code{x}.
}
\description{
Clean a graph object's attributes in a standardized fashion.
}
\examples{
library(snatools)

data("Koenigsberg", package = "igraphdata")

Koenigsberg \%>\% vrt_get_attrs()

Koenigsberg \%>\% strictify() \%>\% strictify()


}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
