% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{vrt_as_df}
\alias{vrt_as_df}
\alias{vrt_as_df.igraph}
\alias{vrt_as_df.network}
\title{Extract all vertex attributes as a data frame.}
\usage{
vrt_as_df(x)

\method{vrt_as_df}{igraph}(x)

\method{vrt_as_df}{network}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\description{
Extract all vertex attributes as a data frame.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
