% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_.R
\name{vrt_attrs}
\alias{vrt_attrs}
\alias{vrt_attrs.igraph}
\alias{vrt_attrs.network}
\title{Extract vertex attributes from a graph.}
\usage{
vrt_attrs(x, enforce_names = TRUE, exclude_na = NULL)

\method{vrt_attrs}{igraph}(ig, enforce_names = TRUE, exclude_na = NULL)

\method{vrt_attrs}{network}(x, enforce_names = TRUE, exclude_na = TRUE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{exclude_na}{}
}
\value{
A named \code{list} of \code{x}'s vertex attributes.
}
\description{
Extract vertex attributes from a graph.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
