% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_delete_attr}
\alias{vrt_delete_attr}
\alias{vrt_delete_attr.bridge_net}
\alias{vrt_delete_attr.igraph}
\alias{vrt_delete_attr.network}
\alias{vrt_delete_attr.tbl_graph}
\title{Delete Vertex Attributes}
\usage{
vrt_delete_attr(x, ...)

\method{vrt_delete_attr}{bridge_net}(x, ...)

\method{vrt_delete_attr}{igraph}(x, ...)

\method{vrt_delete_attr}{network}(x, ...)

\method{vrt_delete_attr}{tbl_graph}(x, ...)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{...}{\code{character} including the name(s) of one or more vertex attributes to delete
from \code{x}.}
}
\value{
\code{x}, but with \code{...} deleted.
}
\description{
Delete Vertex Attributes
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

sampson_monastery \%>\% 
  vrt_delete_attr("cloisterville", "status")
  
ig \%>\% 
  vrt_delete_attr("faction")

nw \%>\% 
  vrt_delete_attr("cloisterville", "faction")

tidy_g \%>\% 
  vrt_delete_attr("faction", "status")

}
\seealso{
\code{\link[igraph:delete_vertex_attr]{igraph::delete_vertex_attr()}}

\code{\link[network:delete.vertex.attribute]{network::delete.vertex.attribute()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
