% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_filter}
\alias{vrt_filter}
\alias{vrt_filter_se}
\alias{vrt_filter.bridge_net}
\alias{vrt_filter_se.bridge_net}
\alias{vrt_filter.igraph}
\alias{vrt_filter_se.igraph}
\alias{vrt_filter.network}
\alias{vrt_filter_se.network}
\title{Filter vertices by their attributes.}
\usage{
vrt_filter(x, ...)

vrt_filter_se(x, vrt_attr, .f, vrt_val = NULL)

\method{vrt_filter}{bridge_net}(x, ...)

\method{vrt_filter_se}{bridge_net}(x, vrt_attr, .f = NULL,
  vrt_val = NULL)

\method{vrt_filter}{igraph}(x, ...)

\method{vrt_filter_se}{igraph}(x, vrt_attr, .f = NULL, vrt_val = NULL)

\method{vrt_filter}{network}(x, ...)

\method{vrt_filter_se}{network}(x, vrt_attr, .f = NULL, vrt_val = NULL)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{...}{\code{logical} conditions used to subset rows when said conditions evaluate to
\code{TRUE}.}

\item{vrt_attr}{\code{character} specifying edge attribute to target. \code{vrt_filter_se()}
only.}

\item{.f}{\code{function} (default: \code{NULL}) specifying the predicate function to apply to
\code{vrt_attr}. \code{vrt_filter_se()} only.}

\item{vrt_val}{(default: \code{NULL}) Value to use as second argument to \code{.f} (if
applicable).}
}
\value{
\code{x}, but with vertices filtered by \code{...} for \code{vrt_filter()} or \code{edge_attr}, \code{.f},
and/or \code{vrt_val} for \code{vrt_filter_se()}.
}
\description{
Filter vertices by their attributes.
}
\details{
\code{vrt_filter()} uses non-standard evaluation to make syntax as convenient as possible,
while \code{vrt_filter_se()} uses standard evaluation if preferred (or if \code{vrt_filter()}
causes trouble).
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network() 

sampson_monastery \%>\% 
  vrt_filter(faction == "Outcasts", status == "Expelled")

sampson_monastery \%>\% 
  vrt_filter_se("faction", `==`, "Outcasts") \%>\% 
  vrt_filter_se("status", `==`, "Expelled")

ig \%>\% 
  vrt_filter(cloisterville, faction == "Loyal")
  
ig \%>\% 
  vrt_filter_se("cloisterville") \%>\% 
  vrt_filter_se("faction", `==`, "Loyal")

nw \%>\% 
  vrt_filter(status \%in\% c("Expelled", "Left Voluntarily"))
  
nw \%>\% 
  vrt_filter_se("status", `\%in\%`, c("Expelled", "Left Voluntarily"))

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
