% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_get_attr}
\alias{vrt_get_attr}
\alias{vrt_get_attr.bridge_net}
\alias{vrt_get_attr.igraph}
\alias{vrt_get_attr.network}
\alias{vrt_get_attr.tbl_graph}
\title{Extract a specific vertex attribute.}
\usage{
vrt_get_attr(x, vrt_attr)

\method{vrt_get_attr}{bridge_net}(x, vrt_attr)

\method{vrt_get_attr}{igraph}(x, vrt_attr)

\method{vrt_get_attr}{network}(x, vrt_attr)

\method{vrt_get_attr}{tbl_graph}(x, vrt_attr)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{vrt_attr}{\code{character} specifying the target vertex attribute.}
}
\value{
\code{vector} of vertex attribute values.
}
\description{
Extract a specific vertex attribute.
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()

tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

vrt_get_attr(sampson_monastery, "faction")

vrt_get_attr(ig, "faction")

vrt_get_attr(nw, "faction")

vrt_get_attr(tidy_g, "faction")

}
\seealso{
\code{\link[=vrt_attr_names]{vrt_attr_names()}}, \code{\link[=vrt_to_df]{vrt_to_df()}}, \code{\link[igraph:vertex_attr]{igraph::vertex_attr()}},
\code{\link[network:get.vertex.attribute]{network::get.vertex.attribute()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
