% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_get_attr_df}
\alias{vrt_get_attr_df}
\alias{vrt_get_attr_df.sna_net}
\alias{vrt_get_attr_df.igraph}
\alias{vrt_get_attr_df.network}
\alias{vrt_get_attr_df.tbl_graph}
\title{Extract all vertex attributes as a data frame.}
\usage{
vrt_get_attr_df(x, sna_vertices_df = TRUE)

\method{vrt_get_attr_df}{sna_net}(x)

\method{vrt_get_attr_df}{igraph}(x, sna_vertices_df = TRUE)

\method{vrt_get_attr_df}{network}(x, sna_vertices_df = TRUE)

\method{vrt_get_attr_df}{tbl_graph}(x, sna_vertices_df = TRUE)
}
\arguments{
\item{x}{An \code{sna_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{sna_vertices_df}{\code{logical} (default: \code{TRUE}) indicating whether to return
an \code{sna_vertices_df} object that standardizes the results of graph objects.}
}
\value{
\code{data.frame} or \code{sna_vertices_df} with all vertex attributes contained in \code{x}.
}
\description{
Extract all vertex attributes as a data frame.
}
\examples{
library(snatools)

adj_mat <- rbinom(16L, 1L, 0.3) \%>\% 
  matrix(nrow = 4L, dimnames = list(letters[1:4], letters[1:4]))

ig <- adj_mat \%>\% 
  igraph::graph_from_adjacency_matrix() \%>\% 
  igraph::set_vertex_attr("a_vertex_attribute", value = sample(seq_len(igraph::vcount(.)))) 

vrt_get_attr_df(ig)
vrt_get_attr_df(ig, sna_vertices_df = FALSE)

nw <- network::as.network(adj_mat)
network::set.vertex.attribute(nw, "a_vertex_attribute", value = sample(seq_len(network::network.size(nw))))

vrt_get_attr_df(nw)
vrt_get_attr_df(nw, sna_vertices_df = FALSE)

tbl_g <- tidygraph::as_tbl_graph(ig)

vrt_get_attr_df(tbl_g)
vrt_get_attr_df(tbl_g, sna_vertices_df = FALSE)

sna_g <- as_sna_net(nw)

vrt_get_attr_df(sna_g)

}
\seealso{
\code{\link[=vrt_get_attr]{vrt_get_attr()}}, \code{\link[=vrt_attr_names]{vrt_attr_names()}}, \code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
