% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_get.R
\name{vrt_get_attrs}
\alias{vrt_get_attrs}
\alias{vrt_get_attrs.igraph}
\alias{vrt_get_attrs.network}
\title{Extract vertex attributes from a graph.}
\usage{
vrt_get_attrs(x)

\method{vrt_get_attrs}{igraph}(ig)

\method{vrt_get_attrs}{network}(nw)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A named \code{list} of \code{x}'s vertex attributes.
}
\description{
Extract vertex attributes from a graph.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
