% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_get_names}
\alias{vrt_get_names}
\alias{vrt_get_names.igraph}
\alias{vrt_get_names.network}
\alias{vrt_get_names.tbl_graph}
\title{Get Vertex Names}
\usage{
vrt_get_names(x)

\method{vrt_get_names}{igraph}(x)

\method{vrt_get_names}{network}(x)

\method{vrt_get_names}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{vector} containing to the values of the vertex attribute that \code{x}'s class uses
for vertex names.
}
\description{
Get Vertex Names
}
\details{
Supported graph objects use the following conventions: \cr
\itemize{
\item \code{igraph} and \code{tbl_graph} objects use \code{name} for vertex names.
\item \code{network} objects use \code{vertex.names} for vertex names.
}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
