% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{vrt_subset}
\alias{vrt_subset}
\title{Subset by Vertices}
\usage{
vrt_subset(x, .predicate)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{.predicate}{A single predicate function describing how \code{x}'s edges should be
subset.}
}
\description{
Subset by Vertices
}
\examples{
library(snatools)

crisis_in_cloister \%>\% 
  vrt_subset(cloisterville)
  
crisis_in_cloister \%>\% 
  as_network() \%>\% 
  vrt_subset(!cloisterville)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
