% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_to_df}
\alias{vrt_to_df}
\alias{vrt_to_df.bridge_net}
\alias{vrt_to_df.igraph}
\alias{vrt_to_df.network}
\alias{vrt_to_df.tbl_graph}
\title{Extract all vertex attributes as a data frame.}
\usage{
vrt_to_df(x, leave_raw = FALSE)

\method{vrt_to_df}{bridge_net}(x)

\method{vrt_to_df}{igraph}(x, leave_raw = FALSE)

\method{vrt_to_df}{network}(x, leave_raw = FALSE)

\method{vrt_to_df}{tbl_graph}(x, leave_raw = FALSE)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{leave_raw}{\code{logical} (default: \code{FALSE}) indicating whether to return
a raw \code{data.frame} instead of a \code{vertex_data_frame} that standardizes results across
graph classes.}
}
\value{
\code{data.frame} or \code{vertex_data_frame} with all vertex attributes contained in \code{x}.
}
\description{
Extract all vertex attributes as a data frame.
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()

tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

vrt_to_df(sampson_monastery)

vrt_to_df(ig)

vrt_to_df(nw)

vrt_to_df(tidy_g)

}
\seealso{
\code{\link[=vrt_get_attr]{vrt_get_attr()}}, \code{\link[=vrt_attr_names]{vrt_attr_names()}}, \code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
