% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.network}
\alias{as_igraph.default}
\alias{as_igraph.igraph}
\alias{as_igraph.ucinet}
\title{Convert \code{network} objects to \code{igraph}.}
\usage{
as_igraph(x, ...)

\method{as_igraph}{network}(x)

\method{as_igraph}{default}(x, ...)

\method{as_igraph}{igraph}(x)

\method{as_igraph}{ucinet}(x, ...)
}
\description{
Convert \code{network} objects to \code{igraph}.
}
