% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.bridge_net}
\alias{as_igraph.igraph}
\alias{as_igraph.network}
\alias{as_igraph.tbl_graph}
\title{Conversion to \code{igraph} objects.}
\usage{
as_igraph(x)

\method{as_igraph}{bridge_net}(x)

\method{as_igraph}{igraph}(x)

\method{as_igraph}{network}(x)

\method{as_igraph}{tbl_graph}(x)
}
\arguments{
\item{x}{\code{\link[network:network]{network::network}} (or \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}}) object.}
}
\value{
An \code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) object.
}
\description{
Accurately map foreign graph data to \code{igraph} objects.
}
\details{
\code{as_igraph()} converts \code{x} to an intermediate \code{bridge_net} object that is capable of
mapping metadata, edges, vertices, and attributes (edge, vertex, and graph-level)
to a new, valid \code{igraph} object. \cr
}
\seealso{
\code{\link[=as_network]{as_network()}}, \code{\link[intergraph:asIgraph]{intergraph::asIgraph()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
