% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.bridge_net}
\alias{as_igraph.igraph}
\alias{as_igraph.network}
\alias{as_igraph.tbl_graph}
\title{Conversion to \code{igraph} objects.}
\usage{
as_igraph(x)

\method{as_igraph}{bridge_net}(x)

\method{as_igraph}{igraph}(x)

\method{as_igraph}{network}(x)

\method{as_igraph}{tbl_graph}(x)
}
\arguments{
\item{x}{\code{\link[network:network]{network::network}} object.}
}
\value{
An \code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) object.
}
\description{
Convert objects of class \code{network} to \code{igraph} objects.
}
\examples{
library(snatools)

data("sampson", package = "ergm")

samplike

samplike \%>\% 
  as_igraph()
  
# bipartite network to igraph =============================================================
southern_women_matrix <- c(
  1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 
  1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
  0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 
  0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1,
  0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1,
  1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0,
  1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
  1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 1,
  1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0,
  1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1,
  1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0
  ) \%>\% 
  matrix(nrow = 18, ncol = 14) \%>\% 
  `rownames<-`(c("EVELYN", "LAURA", "THERESA", "BRENDA", "CHARLOTTE", "FRANCES", 
                 "ELEANOR", "PEARL", "RUTH", "VERNE", "MYRNA", "KATHERINE", "SYLVIA", 
                 "NORA", "HELEN", "DOROTHY", "OLIVIA", "FLORA"))

sw_nw <- network::as.network.matrix(southern_women_matrix)

sw_nw

sw_nw \%>\% 
  as_igraph()
  
# networkDynamic to igraph ================================================================
data("windsurfers", package = "networkDynamic")

windsurfers

windsurfers \%>\% 
  as_igraph()

}
\seealso{
\code{\link[=as_network]{as_network()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
