% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.igraph}
\alias{as_igraph.network_dataset}
\alias{as_igraph.network}
\alias{as_igraph.tbl_graph}
\alias{as_igraph.ucinet}
\title{Conversion to \code{igraph} objects.}
\usage{
as_igraph(x, ...)

\method{as_igraph}{igraph}(x)

\method{as_igraph}{network_dataset}(x)

\method{as_igraph}{network}(x)

\method{as_igraph}{tbl_graph}(x, ...)

\method{as_igraph}{ucinet}(x, ...)
}
\arguments{
\item{x}{\code{\link[network:network]{network::network}} object.}

\item{...}{Additional arguments on to other methods (\code{clean_graph}, \code{actor_type}).}

\item{clean_graph}{logical. Whether to automatically call \code{clean_graph()} before converting x. \cr
Default: \code{TRUE}}

\item{actor_type}{logical. For bipartite graphs, the \code{"type"} (vertex attribute) that
"actor" vertices are to be assigned. \cr
Default: \code{TRUE}}
}
\value{
An \code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) object.
}
\description{
Conversion to \code{igraph} objects.
}
