% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_network.R
\name{as_network}
\alias{as_network}
\alias{as_network.network}
\alias{as_network.network_dataset}
\alias{as_network.igraph}
\alias{as_network.tbl_graph}
\title{Conversion to \code{network} objects.}
\usage{
as_network(x, ...)

\method{as_network}{network}(x, ...)

\method{as_network}{network_dataset}(x)

\method{as_network}{igraph}(x, actor_type = TRUE)

\method{as_network}{tbl_graph}(x, ...)
}
\arguments{
\item{x}{\code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) or \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} object.}

\item{...}{Additional arguments passed on to other methods (currently \code{actor_type}).}

\item{actor_type}{\code{logical}. For bipartite graphs, the vertex \code{"type"} specifying which vertices in
\code{x} are to be treated as "actors" in returned \code{network} object. See Details. \cr}
}
\value{
A \code{\link[network:network]{network::network}} object.
}
\description{
Conversion to \code{network} objects.
}
\details{
\code{actor_type}
\itemize{
\item By default, \code{igraph} represents bipartite graphs using a \code{logical} vertex attribute named \code{type}.
\item \code{network} represents bipartite graphs through a count of the vertices in the "actors"
partition.
\item \code{actor_type} provides a way to specify which vertices of an \code{igraph} object should be
represented as the "actors" in the resulting \code{network} object (as opposed to the vertices
of the second mode, e.g. "events").
}
}
\examples{
library(snatools)

data("sampson_monastery")
  
sampson_monastery \%>\% 
  as_network()
  
sampson_monastery \%>\% 
  as_igraph() \%>\% 
  as_network()

as_igraph(sampson_monastery) \%==\% as_network(sampson_monastery)

data("davis_southern_women")

davis_southern_women \%>\% # bipartite
  as_network()
  
davis_southern_women \%>\% 
  as_igraph() \%>\% 
  as_network()
  
as_network(davis_southern_women) \%==\% as_igraph(davis_southern_women)

}
\seealso{
\code{\link[=as_igraph]{as_igraph()}}, \code{\link[intergraph:asNetwork]{intergraph::asNetwork()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
