% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_network.R
\name{as_network}
\alias{as_network}
\alias{as_network.bridge_net}
\alias{as_network.igraph}
\alias{as_network.network}
\alias{as_network.tbl_graph}
\title{Conversion to \code{network} objects.}
\usage{
as_network(x)

\method{as_network}{bridge_net}(x)

\method{as_network}{igraph}(x)

\method{as_network}{network}(x)

\method{as_network}{tbl_graph}(x)
}
\arguments{
\item{x}{\code{igraph} (\code{\link[igraph:graph]{igraph::graph}}) or \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} object.}
}
\value{
A \code{\link[network:network]{network::network}} object.
}
\description{
Convert objects of class \code{igraph} to \code{network} objects.
}
\examples{
library(snatools)

crisis_in_cloister

crisis_in_cloister \%>\% 
  as_network()
  
florence

florence \%>\% 
  as_network()
  
# bipartite igraph to network =============================================================
southern_women

southern_women \%>\% 
  as_network()

# tidygraph to network ====================================================================
suppressPackageStartupMessages(library(tidygraph))

tg <- crisis_in_cloister \%>\% 
  as_tbl_graph()
  
tg

tg \%>\% 
  as_network()

}
\seealso{
\code{\link[=as_igraph]{as_igraph()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
