% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_network.R
\name{as_network}
\alias{as_network}
\alias{as_network.igraph}
\alias{as_network.default}
\alias{as_network.network}
\title{Convert a \code{igraph} objects to \code{network}.}
\usage{
as_network(x, ...)

\method{as_network}{igraph}(x)

\method{as_network}{default}(x, ...)

\method{as_network}{network}(x)
}
\arguments{
\item{x}{A graph object.\code{igraph} objects are supported.}
}
\value{
A \code{network} object.
}
\description{
Convert a \code{igraph} objects to \code{network}.
}
