% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_attr_names}
\alias{edg_attr_names}
\alias{edg_attr_names.igraph}
\alias{edg_attr_names.network}
\alias{edg_attr_names.tbl_graph}
\title{Get Edge Attribute Names}
\usage{
edg_attr_names(x)

\method{edg_attr_names}{igraph}(x)

\method{edg_attr_names}{network}(x)

\method{edg_attr_names}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{character} \code{vector} of \code{x}'s vertex attribute names or \code{NULL} if
no edge attributes exist.
}
\description{
Get Edge Attribute Names
}
\examples{
library(snatools)

florence \%>\% 
  edg_attr_names()

}
\seealso{
\code{\link[igraph:edge_attr_names]{igraph::edge_attr_names()}}

\code{\link[network:list.edge.attributes]{network::list.edge.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
