% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_attr_names}
\alias{edg_attr_names}
\alias{edg_attr_names.bridge_net}
\alias{edg_attr_names.igraph}
\alias{edg_attr_names.network}
\alias{edg_attr_names.tbl_graph}
\title{Extract a graph objects's edge attribute names.}
\usage{
edg_attr_names(x)

\method{edg_attr_names}{bridge_net}(x)

\method{edg_attr_names}{igraph}(x)

\method{edg_attr_names}{network}(x)

\method{edg_attr_names}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{character} \code{vector} listing the names of \code{x}'s edge attributes.
}
\description{
Extract a graph objects's edge attribute names.
}
\examples{
library(snatools)

adj_mat <- rbinom(16L, 1L, 0.3) \%>\% 
  matrix(nrow = 4L, dimnames = list(letters[1:4], letters[1:4]))

ig <- adj_mat \%>\% 
  igraph::graph_from_adjacency_matrix() \%>\% 
  igraph::set_edge_attr("another_attr", value = sample(seq_len(igraph::ecount(.)))) 
edg_attr_names(ig)

nw <- network::as.network(adj_mat)
network::set.edge.attribute(nw, "another_attr", 
                            value = sample(seq_len(network::network.edgecount(nw))))
edg_attr_names(nw)

tbl_g <- tidygraph::as_tbl_graph(ig)
edg_attr_names(tbl_g)

sna_g <- as_bridge_net(ig)
edg_attr_names(sna_g)

}
\seealso{
\code{\link[=edg_get_attr]{edg_get_attr()}}, \code{\link[=edg_to_df]{edg_to_df()}}, \code{\link[igraph:edge_attr_names]{igraph::edge_attr_names()}},
\code{\link[network:list.edge.attributes]{network::list.edge.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
