% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-attributes.R
\name{edg_get_attr}
\alias{edg_get_attr}
\alias{edg_get_attr.igraph}
\alias{edg_get_attr.network}
\alias{edg_get_attr.tbl_graph}
\title{Get a Specific Edge Attribute}
\usage{
edg_get_attr(x, edg_attr, .simplify = TRUE, .null_as_na = FALSE)

\method{edg_get_attr}{igraph}(x, edg_attr, .simplify = TRUE,
  .null_as_na = FALSE)

\method{edg_get_attr}{network}(x, edg_attr, .simplify = TRUE,
  .null_as_na = FALSE)

\method{edg_get_attr}{tbl_graph}(x, edg_attr, .simplify = TRUE,
  .null_as_na = FALSE)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{edg_attr}{\code{character} specifying the target edge attribute.}
}
\value{
\code{vector} containing the vertex attributes specified by \code{vrt_attr}.
}
\description{
Get a Specific Edge Attribute
}
\examples{
library(snatools)

florence \%>\% 
  edg_get_attr("relation")

}
\seealso{
\code{\link[igraph:edge_attr]{igraph::edge_attr()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
