% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_index.R
\name{ei_index}
\alias{ei_index}
\title{E-I Index}
\usage{
ei_index(x, vrt_attr, scope = c("global", "group", "vertex"))
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{vrt_attr}{\code{character} specifying the target vertex attribute.}

\item{scope}{\code{character} (default: \code{"global"}) indicating the target scope.
Choices: \code{"global"}, \code{"group"}, or \code{"vertex"}.}
}
\description{
Given a categorical vertex attribute describing mutually exclusive groups, the E-I
index represents a ratio of external to internal ties.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
