% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_.R
\name{ei_index}
\alias{ei_index}
\title{E-I Index}
\usage{
ei_index(x, vrt_attr, scope = c("global", "group", "vertex", "full"),
  drop_loops = FALSE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{vrt_attr}{\code{character} representing the name of a categorical vertex attribute present in \code{x}.}

\item{scope}{\code{character}. One of \code{"global"}, \code{"group"}, \code{"vertex"}, or \code{"full"}.}

\item{drop_loops}{\code{logical}. Whether loops should be removed prior to calculations. \cr
Default: \code{FALSE}}
}
\value{
One of:
\itemize{
\item scalar \code{double}
\item \code{ei_index_grp} (\code{data.frame})
\item \code{ei_index_vrt} (\code{data.frame})
\item \code{ei_index_full} (named \code{list}) containing a scalar \code{double}, an \code{ei_index_grp},
and an \code{ei_index_vrt}.
See Details.
}
}
\description{
Given a categorical vertex attribute describing mutually exclusive groups, the E-I
index represents a ratio of external to internal ties.
}
\details{
\deqn{E\mbox{-}I~Index = \frac{EL-IL}{EL+IL}}
\itemize{
\item Variables and  Interpretation:
\itemize{
\item \eqn{EL} is the count of external ties and \eqn{IL} is the count of internal ties.
\item Results closer to \code{-1} suggest tendencies toward homophily while results closer
to \code{+1} suggest tendencies toward heterophily.
\cr\cr
}
\item \code{ei_index_global()} is an alias for \code{ei_index()} with argument \code{scope = "full"}.
\itemize{
\item The global E-I Index calculates the ratio of external ties to internal ties for
\strong{all groups}.
\item Returns scalar \code{double}
}
\item \code{ei_index_grp()} is an alias for \code{ei_index()} with argument \code{scope = "group"}.
\itemize{
\item The group E-I Index calculates the ratio of external to internal ties for \strong{each
group}.
\item Returns \code{ei_index_grp} object, a \code{data.frame} with columns:
\itemize{
\item \code{attribute}. Group category, as present in the \code{vrt_attr} provided.
\item \code{external_ties}. Count of ties external to the group.
\item \code{internal_ties}. Count of ties internal to the group.
\item \code{ei_index}. Value calculated by equation above.
}
}
\item \code{ei_index_vrt()} is an alias for \code{ei_index()} with argument \code{scope = "vertex"}.
\itemize{
\item The vertex E-I Index calculates the ratio of external to internal ties for \strong{each
vertex}.
\item Returns \code{ei_index_vrt} object, a \code{data.frame} with columns:
\itemize{
\item \code{name}. Vertex names, as found in \code{x}'s \code{name} or \code{vertex.names} vertex attribute.
\item \code{attribute}. Group to which vertices belong, as present in the \code{vrt_attr} provided.
\item \code{external_ties}. Count of ties external to the group.
\item \code{internal_ties}. Count of ties internal to the group.
\item \code{ei_index}. Value calculated by equation above.
}
}
\item \code{ei_index()} with argument \code{scope = "full"} calculates global, group, and vertex E-I
Indices.
\itemize{
\item Returns \code{ei_index_full} object, a \code{list} with elements (in formats described above):
\itemize{
\item \code{global}
\item \code{by_group}
\item \code{by_vertex}
}
}
}
}
\examples{
library(snatools)

data("samplk", package = "ergm")

ei_index(samplk1, vrt_attr = "group", scope = "full")

(ei_global <- ei_index_global(samplk1, vrt_attr = "group"))

(ei_by_group <- ei_index_grp(samplk1, vrt_attr = "group"))

autoplot(ei_by_group)

(ei_by_vertex <- ei_index_vrt(samplk1, vrt_attr = "group"))

autoplot(ei_by_vertex)

}
\references{
Krackhardt, David, and Robert N. Stern. "Informal Networks and
Organizational Crises: An Experimental Simulation." Social Psychology Quarterly 51, no.
2 (1988): 123-40. \url{http://www.jstor.org/stable/2786835}.
}
\seealso{
\code{\link[=ei_index_global_permute]{ei_index_global_permute()}}, \code{\link[=rep_mixing_matrix]{rep_mixing_matrix()}}, \code{\link[network:mixingmatrix]{network::mixingmatrix()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
