% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_index.R
\name{ei_index}
\alias{ei_index}
\alias{print.ei_global}
\alias{autoplot.ei_group}
\title{E-I Index}
\usage{
ei_index(x, vrt_attr, scope = c("global", "group", "vertex"),
  drop_loops = TRUE)

\method{print}{ei_global}(x, ...)

\method{autoplot}{ei_group}(object, ...)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{vrt_attr}{\code{character} representing the name of a categorical vertex attribute present in \code{x}.}

\item{scope}{\code{character}. One of \code{"global"}, \code{"group"}, \code{"vertex"}, or \code{"full"}.}

\item{drop_loops}{\code{logical} (default: \code{TRUE}) indicating whether loops should be
removed prior to calculations.}
}
\description{
Given a categorical vertex attribute describing mutually exclusive groups, the E-I
index represents a ratio of external to internal ties.
}
\details{
\deqn{E\mbox{-}I~Index = \frac{EL-IL}{EL+IL}}
}
\examples{
library(snatools)

data(sampson_monastery)

sampson_monastery \%>\% 
  edg_filter(time == 1) \%>\% 
  ei_index(vrt_attr = "faction", scope = "global")
  
sampson_monastery \%>\% 
  edg_filter(time == 3, positive_relation) \%>\% 
  ei_index(vrt_attr = "faction", scope = "group") \%>\% 
  autoplot()
  
}
\references{
Krackhardt, David, and Robert N. Stern. "Informal Networks and
Organizational Crises: An Experimental Simulation." Social Psychology Quarterly 51, no.
2 (1988): 123-40. \url{http://www.jstor.org/stable/2786835}.
}
\seealso{
\code{\link[=ei_index_global_permute]{ei_index_global_permute()}}, \code{\link[=rep_mixing_matrix]{rep_mixing_matrix()}}, \code{\link[network:mixingmatrix]{network::mixingmatrix()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
