% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edg_get_.R
\name{edg_get_attrs}
\alias{edg_get_attrs}
\alias{edg_get_attrs.igraph}
\alias{edg_get_attrs.network}
\alias{edg_get_attr}
\alias{edg_get_attr.igraph}
\alias{edg_get_attr.network}
\alias{edg_get_attr_names}
\alias{edg_get_attr_names.igraph}
\alias{edg_get_attr_names.network}
\title{Extract Edge Attributes}
\usage{
edg_get_attrs(x, ...)

\method{edg_get_attrs}{igraph}(x, include_el = FALSE)

\method{edg_get_attrs}{network}(x, include_el = FALSE,
  ignore_na = TRUE)

edg_get_attr(x, edg_attr, ...)

\method{edg_get_attr}{igraph}(x, edg_attr, ...)

\method{edg_get_attr}{network}(x, edg_attr, ignore_na = TRUE, ...)

edg_get_attr_names(x, ...)

\method{edg_get_attr_names}{igraph}(x)

\method{edg_get_attr_names}{network}(x, ignore_na = TRUE)
}
\arguments{
\item{x}{A graph object.}

\item{...}{Arguments passed on to other methods. See below.}

\item{include_el}{\code{logical}. Whether to include edge list columns indicating which
vertices are part of each edge's dyad.}

\item{ignore_na}{\code{logical}. Whether to ignore the \code{"na"} attribute of
\code{\link[network:network]{network::network}} objects. \cr
Default: \code{TRUE}}

\item{edg_attr}{\code{character}. Name of target attribute.
Whether to attempt calling \code{\link[tibble:as_tibble]{tibble::as_tibble()}} in \code{edg_get_attrs_df()}. \cr
Default: \code{TRUE}}
}
\value{
A named \code{list}, \code{vector}, or \code{\link[tibble:tibble]{tibble::tibble}}. See Details.
}
\description{
Extract edge attributes in a variety of formats consistent across
\code{igraph} and \code{network} objects.
}
\details{
\itemize{
\item \code{edg_get_attrs} extract all edge attributes as a named \code{tibble::tibble}.
}

\itemize{
\item \code{edg_get_attr()} extracts the single edge attribute specified by \code{edg_attr} as a \code{vector}.
}

\itemize{
\item \code{edg_get_attr_names()} extract the names of vertex attributes as a \code{character} \code{vector}.
\itemize{
\item \code{edg_get_attr_names()} differs from \code{network::list.edge.attributes()}, which
sorts attribute names alphabetically.
}
}
}
\examples{
library(snatools)

data("enron", package = "igraphdata")
(ig <- enron)

data("sampson", package = "ergm")
(nw <- samplike)

edg_get_attrs(ig, include_el = TRUE)
edg_get_attrs(nw, include_el = TRUE)

edg_get_attr(ig, "Time") \%>\% head(20)
edg_get_attr(nw, "nominations") \%>\% head(20)

}
\seealso{
\code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}

\code{\link[igraph:edge_attr]{igraph::edge_attr()}}, \code{\link[network:get.edge.attribute]{network::get.edge.attribute()}}

\code{\link[igraph:edge_attr_names]{igraph::edge_attr_names()}}, \code{\link[network:list.edge.attributes]{network::list.edge.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
