% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edg_get_.R
\name{edg_get_attrs}
\alias{edg_get_attrs}
\alias{edg_get_attrs.igraph}
\alias{edg_get_attrs.network}
\alias{edg_get_attrs_df}
\alias{edg_get_attr}
\alias{edg_get_attr.network}
\alias{edg_get_attr_names}
\title{Edge Attributes}
\usage{
edg_get_attrs(x, ...)

\method{edg_get_attrs}{igraph}(x)

\method{edg_get_attrs}{network}(x, ignore_na = TRUE)

edg_get_attrs_df(x, try_tibble = TRUE, ...)

edg_get_attr(x, edg_attr, ...)

\method{edg_get_attr}{network}(x, edg_attr, ignore_na = TRUE, ...)

edg_get_attr_names(x, ignore_na = TRUE)
}
\arguments{
\item{x}{A graph object.}

\item{...}{Arguments passed on to other methods. See below.}

\item{ignore_na}{\code{logical}. Whether to ignore the \code{"na"} attribute of
\code{\link[network:network]{network::network}} objects. \cr
Default: \code{TRUE}}

\item{try_tibble}{\code{logical}.
Whether to attempt calling \code{\link[tibble:as_tibble]{tibble::as_tibble()}} in \code{edg_get_attrs_df()}. \cr
Default: \code{TRUE}}

\item{edg_attr}{\code{character}. Name of target attribute.}
}
\value{
A named \code{list}, \code{vector}, \code{data.frame}, or \code{\link[tibble:tibble]{tibble::tibble}}. See Details.
}
\description{
Extract edge attributes in a variety of formats.
}
\details{
\itemize{
\item \code{edg_get_attrs} extract all edge attributes as a named \code{list}.
}

\itemize{
\item \code{edg_get_attrs_df} is a convenience wrapper around \code{edg_get_attrs()} that returns a
\code{data.frame}.
\itemize{
\item \code{stringsAsFactors} is \emph{always} \code{FALSE}.
\item If available, \code{\link[tibble:as_tibble]{tibble::as_tibble()}} is called to return a \code{tbl_df} instead.
}
}

\itemize{
\item \code{edg_get_attr} extracts the single edge attribute specified by \code{edg_attr} as a \code{vector}.
}

\itemize{
\item \code{edg_get_attr_names} extract the names of vertex attributes as a \code{character} \code{vector}.
}
}
\examples{
library(snatools)

data("enron", package = "igraphdata")
(ig <- enron)

data("sampson", package = "ergm")
(nw <- samplike)

edg_get_attr_names(ig)
edg_get_attr_names(nw)

edg_get_attrs(ig) \%>\% lapply(head)
edg_get_attrs(nw)

edg_get_attrs_df(ig)
edg_get_attrs_df(nw)

edg_get_attr(ig, edg_attr = "Time") \%>\% head()
edg_get_attr(nw, edg_attr = "nominations")

}
\seealso{
\code{\link[igraph:edge_attr]{igraph::edge_attr()}}, \code{\link[igraph:edge_attr_names]{igraph::edge_attr_names()}}, \code{\link[igraph:as_data_frame]{igraph::as_data_frame()}},
\code{\link[network:get.edge.attribute]{network::get.edge.attribute()}}, \code{\link[network:list.edge.attributes]{network::list.edge.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
