% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_get_.R
\name{vrt_get_attrs}
\alias{vrt_get_attrs}
\alias{vrt_get_attrs.igraph}
\alias{vrt_get_attrs.network}
\alias{vrt_get_attrs_df}
\alias{vrt_get_attrs_df.igraph}
\alias{vrt_get_attrs_df.network}
\alias{vrt_get_attr}
\alias{vrt_get_attr.igraph}
\alias{vrt_get_attr.network}
\alias{vrt_get_names}
\alias{vrt_get_names.igraph}
\alias{vrt_get_names.network}
\alias{vrt_get_attr_names}
\alias{vrt_get_attr_names.igraph}
\alias{vrt_get_attr_names.network}
\title{Extract Vertex Attributes}
\usage{
vrt_get_attrs(x, ...)

\method{vrt_get_attrs}{igraph}(x)

\method{vrt_get_attrs}{network}(x, ignore_na = TRUE)

vrt_get_attrs_df(x, ...)

\method{vrt_get_attrs_df}{igraph}(x, try_tibble = TRUE, ...)

\method{vrt_get_attrs_df}{network}(x, try_tibble = TRUE, ...)

vrt_get_attr(x, vrt_attr)

\method{vrt_get_attr}{igraph}(x, vrt_attr)

\method{vrt_get_attr}{network}(x, vrt_attr)

vrt_get_names(x)

\method{vrt_get_names}{igraph}(x)

\method{vrt_get_names}{network}(x)

vrt_get_attr_names(x, ...)

\method{vrt_get_attr_names}{igraph}(x)

\method{vrt_get_attr_names}{network}(x, ignore_na = TRUE)
}
\arguments{
\item{x}{A graph object.}

\item{...}{Arguments passed on to other methods. See below.}

\item{ignore_na}{\code{logical}. Whether to ignore the \code{"na"} attribute of
\code{\link[network:network]{network::network}} objects. \cr
Default: \code{TRUE}}

\item{try_tibble}{\code{logical}.
Whether to attempt calling \code{\link[tibble:as_tibble]{tibble::as_tibble()}} in \code{vrt_get_attrs_df()}. \cr
Default: \code{TRUE}}

\item{vrt_attr}{\code{character}. Name of target attribute.}
}
\value{
A named \code{list}, \code{vector}, \code{data.frame}, or \code{\link[tibble:tibble]{tibble::tibble}}. See Details.
}
\description{
Extract vertex attributes in a variety of formats consistent across
\code{igraph} and \code{network} objects.
}
\details{
\itemize{
\item \code{vrt_get_attrs()} extracts all vertex attributes as a named \code{list}.
}

\itemize{
\item \code{vrt_get_attrs_df()} is a convenience wrapper around \code{vrt_get_attrs()} that returns a
\code{data.frame}.
\itemize{
\item \code{stringsAsFactors} is \emph{always} \code{FALSE}.
\item If \code{try_tibble} is \code{TRUE} and the \code{tibble} package is available,
a \code{\link[tibble:tibble]{tibble::tibble}} is returned.
}
}

\itemize{
\item \code{vrt_get_attr()} extracts the single vertex attribute specified by \code{vrt_attr} as a \code{vector}.
}

\itemize{
\item \code{vrt_get_names()} is a convenience wrapper around \code{vrt_get_attr()} to extract the names
of vertices as a \code{vector}.
\itemize{
\item It assumes \code{igraph} objects use \code{"name"} and \code{network} objects use
\code{"vertex.names"}.
\item If a vertex attribute following this convention is not present, \code{vrt_get_names()}
throws a warning and returns \code{NULL}.
}
}

\itemize{
\item \code{vrt_get_attr_names()} extract the names of vertex attributes as a \code{character} \code{vector}.
\itemize{
\item \code{vrt_get_attr_names()} differs from \code{\link[network:list.vertex.attributes]{network::list.vertex.attributes()}}, which sorts
attribute names alphabetically.
}
}
}
\examples{
library(snatools)

data("karate", package = "igraphdata")
(ig <- karate)

data("emon", package = "network")
(nw <- emon$Cheyenne)

vrt_get_attrs(ig)
vrt_get_attrs(nw)

vrt_get_attrs_df(ig)
vrt_get_attrs_df(nw)

vrt_get_attr(ig, vrt_attr = "Faction")
vrt_get_attr(nw, vrt_attr = "Sponsorship")

vrt_get_names(ig)
vrt_get_names(nw)

vrt_get_attr_names(ig)
vrt_get_attr_names(nw)

}
\seealso{
\code{\link[igraph:vertex_attr]{igraph::vertex_attr()}}

\code{\link[igraph:as_data_frame]{igraph::as_data_frame()}}, \code{\link[tibble:as_tibble]{tibble::as_tibble()}}

\code{\link[igraph:vertex_attr]{igraph::vertex_attr()}}, \code{\link[network:get.vertex.attribute]{network::get.vertex.attribute()}}

\code{\link[igraph:vertex_attr_names]{igraph::vertex_attr_names()}}, \code{\link[network:list.vertex.attributes]{network::list.vertex.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
