% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{florence}
\alias{florence}
\title{Padgett's Florentine Families}
\format{\code{igraph}}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm}
}
\usage{
florence
}
\description{
Padgett's Florentine Families
}
\details{
\strong{Background}, \emph{as described in UCINET IV's manual}
\itemize{
\item Breiger & Pattison (1986), in their discussion of local role analysis, use a
subset of data on the social relations among Renaissance Florentine families
(person aggregates) collected by John Padgett from historical documents. The two
relations are business ties (PADGB - specifically, recorded financial ties such as
loans, credits and joint partnerships) and marriage alliances (PADGM).
\item As Breiger & Pattison point out, the original data are symmetrically coded. This is
acceptable perhaps for marital ties, but is unfortunate for the financial ties (which
are almost certainly directed). To remedy this, the financial ties can be recoded as
directed relations using some external measure of power - for instance, a measure of
wealth. PADGW provides information on (1) each family's net wealth in 1427 (in
thousands of lira); (2) the number of priorates (seats on the civic council) held
between 1282- 1344; and (3) the total number of business or marriage ties in the
total dataset of 116 families (see Breiger & Pattison (1986), p 239).
}
}
\examples{
library(snatools)
library(ggraph, quietly = TRUE)

data("florence")

florence \%>\% 
  ggraph(layout = "fr") +
  geom_edge_fan(aes(color = type)) +
  geom_node_point() +
  geom_node_text(aes(label = name), repel = TRUE) +
  theme_void()

}
\references{
\itemize{
\item Breiger R. and Pattison P. (1986). Cumulated social roles: The duality of persons and
their algebras. Social Networks, 8, 215-256.
\item Kent D. (1978). The rise of the Medici: Faction in Florence, 1426-1434. Oxford:
Oxford University Press.
}
}
\keyword{datasets}
