% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%==\%}
\alias{\%==\%}
\title{Test if objects are \code{identical}.}
\usage{
lhs \%==\% rhs
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand side}
}
\value{
\code{logical} Whether \code{lhs} and \code{rhs} are \code{identical()}.
}
\description{
This infix operator provides a stricter method of comparing objects while preventing
surprise type conversions.
}
\examples{

TRUE == 1L
TRUE == 1.0000000000000001
TRUE \%==\% 1L
TRUE \%==\% 1.0000000000000001

NA == NA_character_
NA \%==\% NA_character_

NULL == NA_integer_
NULL \%==\% NA_integer_

NA == "NA"
NA \%==\% "NA"

}
\seealso{
\code{\link[base:identical]{base::identical()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
