% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_.R
\name{mix_ei_index}
\alias{mix_ei_index}
\title{Krackhardt's E-I Index}
\usage{
mix_ei_index(x, vrt_attr)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{vrt_attr}{\code{character} representing a vertex attribute}
}
\description{
Krackhardt's E-I Index
}
