% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-attributes.R
\name{net_attr_names}
\alias{net_attr_names}
\alias{net_attr_names.bridge_net}
\alias{net_attr_names.igraph}
\alias{net_attr_names.network}
\alias{net_attr_names.tbl_graph}
\title{Extract non-structural network attribute names.}
\usage{
net_attr_names(x)

\method{net_attr_names}{bridge_net}(x)

\method{net_attr_names}{igraph}(x)

\method{net_attr_names}{network}(x)

\method{net_attr_names}{tbl_graph}(x)
}
\arguments{
\item{x}{A \code{bridge_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{character} \code{vector} listing the names of \code{x}'s non-structural network
attributes.
}
\description{
Extract non-structural network attribute names.
}
\examples{
library(snatools)

data(sampson_monastery)

ig <- sampson_monastery \%>\% 
  as_igraph()
  
nw <- sampson_monastery \%>\% 
  as_network()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()

net_attr_names(sampson_monastery)

net_attr_names(ig)

net_attr_names(nw)

net_attr_names(tidy_g)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
