% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_count_actors}
\alias{net_count_actors}
\alias{net_count_actors.igraph}
\alias{net_count_actors.network}
\alias{net_count_actors.tbl_graph}
\title{How Many Vertices Belong to a Bipartite Graph's Actor Mode?}
\usage{
net_count_actors(x)

\method{net_count_actors}{igraph}(x)

\method{net_count_actors}{network}(x)

\method{net_count_actors}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{integer} indicating the number of \code{x}'s actor vertices.
}
\description{
How Many Vertices Belong to a Bipartite Graph's Actor Mode?
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
