% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_count_actors}
\alias{net_count_actors}
\alias{net_count_actors.igraph}
\alias{net_count_actors.network}
\alias{net_count_actors.sna_net}
\alias{net_count_actors.tbl_graph}
\title{Count the number of vertices belonging to the actor mode of a bipartite graph object.}
\usage{
net_count_actors(x)

\method{net_count_actors}{igraph}(x)

\method{net_count_actors}{network}(x)

\method{net_count_actors}{sna_net}(x)

\method{net_count_actors}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{sna_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{integer} corresponding to the number \code{x}'s actors.
}
\description{
Count the number of vertices belonging to the actor mode of a bipartite graph object.
}
\examples{
library(snatools)

ig <- igraph::bipartite.random.game(6L, 10L, type = "gnp", 0.4)

net_count_actors(ig)

ig \%>\% 
  as_network() \%>\% 
  net_count_actors()

ig \%>\% 
  as_sna_net() \%>\% 
  net_count_actors() 
  
ig \%>\% 
  tidygraph::as_tbl_graph() \%>\% 
  net_count_actors()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
