% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_count_edges}
\alias{net_count_edges}
\alias{net_count_edges.igraph}
\alias{net_count_edges.network}
\alias{net_count_edges.tbl_graph}
\title{How Many Edges Does a Graph Have?}
\usage{
net_count_edges(x)

\method{net_count_edges}{igraph}(x)

\method{net_count_edges}{network}(x)

\method{net_count_edges}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{integer} indicating the total number of \code{x}'s edges.
}
\description{
How Many Edges Does a Graph Have?
}
\seealso{
\code{\link[igraph:ecount]{igraph::ecount()}}

\code{\link[network:network.edgecount]{network::network.edgecount()}}

\code{\link[tidygraph:graph_size]{tidygraph::graph_size()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
