% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_count_vertices}
\alias{net_count_vertices}
\alias{net_count_vertices.igraph}
\alias{net_count_vertices.network}
\alias{net_count_vertices.tbl_graph}
\title{How Many Vertices Does a Graph Have?}
\usage{
net_count_vertices(x)

\method{net_count_vertices}{igraph}(x)

\method{net_count_vertices}{network}(x)

\method{net_count_vertices}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{integer} indicating the total number of \code{x}'s vertices.
}
\description{
How Many Vertices Does a Graph Have?
}
\seealso{
\code{\link[igraph:vcount]{igraph::vcount()}}

\code{\link[network:network.size]{network::network.size()}}

\code{\link[tidygraph:graph_order]{tidygraph::graph_order()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
