% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_has_isolates}
\alias{net_has_isolates}
\alias{net_has_isolates.igraph}
\alias{net_has_isolates.network}
\alias{net_has_isolates.tbl_graph}
\title{Does a Graph Have Isolates?}
\usage{
net_has_isolates(x)

\method{net_has_isolates}{igraph}(x)

\method{net_has_isolates}{network}(x)

\method{net_has_isolates}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{logical} indicating whether \code{x} is contains isolated vertices.
}
\description{
Does a Graph Have Isolates?
}
\seealso{
\code{\link[tidygraph:node_is_isolated]{tidygraph::node_is_isolated()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
