% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_is_bipartite}
\alias{net_is_bipartite}
\alias{net_is_bipartite.bridge_net}
\alias{net_is_bipartite.igraph}
\alias{net_is_bipartite.network}
\alias{net_is_bipartite.tbl_graph}
\alias{net_is_bipartite.edgelist}
\title{Is a graph object bipartite?}
\usage{
net_is_bipartite(x)

\method{net_is_bipartite}{bridge_net}(x)

\method{net_is_bipartite}{igraph}(x)

\method{net_is_bipartite}{network}(x)

\method{net_is_bipartite}{tbl_graph}(x)

\method{net_is_bipartite}{edgelist}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{logical} indicating whether \code{x} is bipartite.
}
\description{
Is a graph object bipartite?
}
\examples{
library(snatools)

ig <- igraph::bipartite.random.game(7, 3, type = "gnp", p = 0.4)

nw <- ig \%>\% 
  as_network()
  
bridge_g <- ig \%>\% 
  as_bridge_net()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()
  
el <- ig \%>\% 
  rep_as_edgelist(use_names = FALSE)

net_is_bipartite(bridge_g)

net_is_bipartite(ig)

net_is_bipartite(nw)

net_is_bipartite(tidy_g)

net_is_bipartite(el)

}
\seealso{
\code{\link[igraph:is_bipartite]{igraph::is_bipartite()}}

\code{\link[network:is.bipartite]{network::is.bipartite()}}

\code{\link[tidygraph:graph_is_bipartite]{tidygraph::graph_is_bipartite()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
