% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-net.R
\name{net_is_directed}
\alias{net_is_directed}
\alias{net_is_directed.igraph}
\alias{net_is_directed.network}
\title{Is an Object Directed?}
\usage{
net_is_directed(x, ...)

\method{net_is_directed}{igraph}(x, ...)

\method{net_is_directed}{network}(x, ...)
}
\arguments{
\item{x}{\code{<igraph>} or \code{<network>}}
}
\value{
\code{<logical>}
}
\description{
Is an Object Directed?
}
\examples{
ig <- example_igraph()
net_is_directed(ig)

nw <- example_network()
net_is_directed(nw)

}
\seealso{
\code{\link[igraph:is_directed]{igraph::is_directed()}}, \code{\link[network:is.directed]{network::is.directed()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
