% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metadata.R
\name{net_is_directed}
\alias{net_is_directed}
\alias{net_is_directed.bridge_net}
\alias{net_is_directed.igraph}
\alias{net_is_directed.network}
\alias{net_is_directed.tbl_graph}
\alias{net_is_directed.edgelist}
\alias{net_is_directed.adj_matrix}
\title{Are a network's edges directed or undirected?}
\usage{
net_is_directed(x)

\method{net_is_directed}{bridge_net}(x)

\method{net_is_directed}{igraph}(x)

\method{net_is_directed}{network}(x)

\method{net_is_directed}{tbl_graph}(x)

\method{net_is_directed}{edgelist}(x)

\method{net_is_directed}{adj_matrix}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}
}
\value{
\code{logical} indicating whether \code{x}'s edges are directed.
}
\description{
Are a network's edges directed or undirected?
}
\examples{
library(snatools)

ig <- igraph::random.graph.game(10, 0.25, directed = TRUE)

nw <- ig \%>\% 
  as_network()
  
bridge_g <- ig \%>\% 
  as_bridge_net()
  
tidy_g <- ig \%>\% 
  tidygraph::as_tbl_graph()
  
el <- ig \%>\% 
  rep_as_edgelist(use_names = FALSE)
  
adj_mat <- ig \%>\% 
  rep_as_adj_matrix()

net_is_directed(bridge_g)

net_is_directed(ig)

net_is_directed(nw)

net_is_directed(tidy_g)

net_is_directed(el)

net_is_directed(adj_mat)

}
\seealso{
\code{\link[igraph:is_directed]{igraph::is_directed()}}

\code{\link[network:is.directed]{network::is.directed()}}

\code{\link[tidygraph:graph_is_directed]{tidygraph::graph_is_directed()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
