% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{net_is_multiplex}
\alias{net_is_multiplex}
\alias{net_is_multiplex.igraph}
\alias{net_is_multiplex.network}
\alias{net_is_multiplex.tbl_graph}
\title{Is a Graph Multiplex?}
\usage{
net_is_multiplex(x)

\method{net_is_multiplex}{igraph}(x)

\method{net_is_multiplex}{network}(x)

\method{net_is_multiplex}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{logical} indicating whether \code{x} is contains parallel edges.
}
\description{
Is a Graph Multiplex?
}
\seealso{
\code{\link[igraph:any_multiple]{igraph::any_multiple()}}

\code{\link[network:is.multiplex]{network::is.multiplex()}}

\code{\link[tidygraph:graph_is_simple]{tidygraph::graph_is_simple()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
