% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{node_count}
\alias{node_count}
\alias{node_count.igraph}
\alias{node_count.network}
\title{Count the number of nodes in a network.}
\usage{
node_count(.net)

\method{node_count}{igraph}(.net)

\method{node_count}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{integer} scalar
}
\description{
Count the number of nodes in a network.
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
