% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{node_get_attr}
\alias{node_get_attr}
\alias{node_get_attr.igraph}
\alias{node_get_attr.network}
\title{Extract a node attribute.}
\usage{
node_get_attr(.net, .node_attr, .node_index = node_seq(.net))

\method{node_get_attr}{igraph}(.net, .node_attr,
  .node_index = node_seq(.net))

\method{node_get_attr}{network}(.net, .node_attr,
  .node_index = node_seq(.net))
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}

\item{.node_attr}{\code{character} scalar corresponding to the name of a node attribute in \code{.net}.}

\item{.node_index}{\code{numeric} or \code{logical} vector indicating which nodes to target.}
}
\value{
\code{vector} (typically)
}
\description{
Extract a node attribute.
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
