% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{node_get_names}
\alias{node_get_names}
\alias{node_get_names.igraph}
\alias{node_get_names.network}
\title{Extract node names.}
\usage{
node_get_names(.net)

\method{node_get_names}{igraph}(.net)

\method{node_get_names}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{vector} corresponding to node names. See Details.
}
\description{
Extract node names.
}
\details{
\itemize{
\item \code{igraph} uses \code{"name"} to name its vertices. \code{node_get_names()} returns the \code{"name"}
attribute if present. If not, \code{node_get_names()} returns \code{node_seq(.net)}
\item \code{network} uses \code{"vertex.names"} to names it vertices. \code{node_get_names()} always
returns the \code{"vertex.names"} attribute.
}
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
