% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ucinet.R
\name{read_ucinet}
\alias{read_ucinet}
\title{Read UCINET's .dl files}
\usage{
read_ucinet(path, directed = NULL)
}
\arguments{
\item{path}{Path to file or URL.}
}
\value{
A \code{net_primitive} object.
}
\description{
Read UCINET's .dl files
}
\examples{
library(snatools)

"http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/sampson.dat" \%>\%
  read_ucinet()
  
"http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/davis.dat" \%>\% 
  read_ucinet() \%>\% 
  as_igraph()
  
"http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/newfrat.dat" \%>\% 
  read_ucinet() \%>\% 
  as_network()

}
