% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_adjacency_matrix}
\alias{rep_adjacency_matrix}
\alias{rep_adjacency_matrix.igraph}
\alias{rep_adjacency_matrix.network}
\alias{rep_attr_adj_mat}
\alias{rep_attr_adj_mat.igraph}
\alias{rep_attr_adj_mat.network}
\title{Build a network's adjacency matrix representation}
\usage{
rep_adjacency_matrix(x, ...)

\method{rep_adjacency_matrix}{igraph}(x, edg_attr = NULL,
  sparse = FALSE, ...)

\method{rep_adjacency_matrix}{network}(x, edg_attr = NULL, ...)

rep_attr_adj_mat(x, vrt_attr)

\method{rep_attr_adj_mat}{igraph}(x, vrt_attr)

\method{rep_attr_adj_mat}{network}(x, vrt_attr)
}
\arguments{
\item{...}{}
}
\value{
\code{matrix} or \code{dgCMatrix}
}
\description{
Build a network's adjacency matrix representation
}
\details{
\code{rep_attr_adj_mat()} creates an adjacency matrix replacing vertex names with the category
in \code{vrt_attr} to which each vertex belongs.
}
\examples{
library(snatools)

data("sampson", package = "ergm")

rep_adjacency_matrix(samplike)

}
