% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency-matrix.R
\name{rep_as_adj_matrix}
\alias{rep_as_adj_matrix}
\alias{print.adj_matrix}
\title{Construct a graph object's adjacency matrix representation.}
\usage{
rep_as_adj_matrix(x, use_names = TRUE, vrt_attr = NULL,
  sparse = TRUE)

\method{print}{adj_matrix}(x)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, or\code{tbl_graph}, or \code{edgelist} object.}

\item{use_names}{\code{logical} (default: \code{TRUE}) indicating whether to use vertex names for
row and column names.}

\item{vrt_attr}{\code{character} (default: \code{NULL}) indicating which vertex attribute to use
in for row and column names. If provided, \code{vrt_attr} overrides \code{use_names}.`}

\item{sparse}{\code{logical} (default: \code{TRUE}) indicating whether to use
\code{Matrix::sparseMatrix()} in constructing the adjacency matrix.}
}
\value{
An \code{adj_matrix} object.
}
\description{
Construct a graph object's adjacency matrix representation.
}
\examples{
library(snatools)

ig <- igraph::graph("Krackhardt Kite") \%>\% 
  igraph::set_vertex_attr("name", value = paste0("vrt_",
                                                  seq_len(igraph::vcount(.)))) \%>\% 
  igraph::set_vertex_attr("group", value = sample(letters, igraph::vcount(.))) \%>\% 
  igraph::set_vertex_attr("group", value = sample(letters, igraph::vcount(.)))
  
rep_as_adj_matrix(ig)

rep_as_adj_matrix(ig, vrt_attr = "group")
  
nw <- as_network(ig)

rep_as_adj_matrix(nw, use_names = FALSE)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
