% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_matrix.R
\name{rep_as_adj_matrix}
\alias{rep_as_adj_matrix}
\title{Construct a graph object's edgelist representation.}
\usage{
rep_as_adj_matrix(x, use_names = TRUE, vrt_attr = NULL,
  edg_attr = NULL)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{use_names}{\code{logical} (default: \code{TRUE}) indicating whether to use vertex names as
the values filling the returned object. If \code{FALSE}, or a relevant attribute does not
exist, the returned object uses vertex indices.}

\item{vrt_attr}{\code{character} specifying the target vertex attribute.}

\item{edg_attr}{\code{character} (default: \code{NULL}) indicating which edge attribute to use
as fill values in the returned object.}
}
\value{
\code{matrix} with \code{net_count_vertices(x)} rows and columns
}
\description{
Construct a graph object's edgelist representation.
}
\examples{
library(snatools)

florence \%>\% 
  edg_subset(relation == "marriage") \%>\% 
  rep_as_adj_matrix() \%>\% 
  `colnames<-`(rep("", ncol(.)))
  
crisis_in_cloister \%>\% 
  edg_subset(time == 1 & affect == 3) \%>\% 
  rep_as_adj_matrix() \%>\% 
  `colnames<-`(rep("", ncol(.)))
  
florence \%>\% 
  rep_as_adj_matrix(edg_attr = "relation") \%>\% 
  `colnames<-`(rep("", ncol(.)))

}
\seealso{
\code{\link[igraph:as_adjacency_matrix]{igraph::as_adjacency_matrix()}}, \code{\link[network:as.matrix.network.adjacency]{network::as.matrix.network.adjacency()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
