% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{rep_as_edgelist}
\alias{rep_as_edgelist}
\title{Construct a graph object's edgelist representation.}
\usage{
rep_as_edgelist(x, use_names = TRUE, vrt_attr = NULL)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{use_names}{\code{logical} (default: \code{TRUE}) indicating whether to use vertex names as
the values filling the returned object. If \code{FALSE}, or a relevant attribute does not
exist, the returned object uses vertex indices.}

\item{vrt_attr}{\code{character} (default: \code{NULL}) indicating which vertex attribute to use
as fill values in the returned object. If provided, \code{vrt_attr} overrides \code{use_names}.}
}
\value{
\code{matrix} with \code{net_count_edges(x)} rows and 2 columns (\code{.ego} and \code{.alter}).
}
\description{
Construct a graph object's edgelist representation.
}
\examples{
library(snatools)

crisis_in_cloister \%>\% 
  rep_as_edgelist() \%>\% 
  head()
  
crisis_in_cloister \%>\% 
  rep_as_edgelist(use_names = FALSE) \%>\% 
  head()
  
crisis_in_cloister \%>\% 
  rep_as_edgelist(vrt_attr = "cloisterville") \%>\% 
  head()

}
\seealso{
\code{\link[igraph:as_edgelist]{igraph::as_edgelist()}}, \code{\link[network:as.matrix.network.edgelist]{network::as.matrix.network.edgelist()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
