% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-matrix.R
\name{rep_as_mixing_matrix}
\alias{rep_as_mixing_matrix}
\alias{is_mixing_matrix}
\alias{print.mixing_matrix}
\alias{as.matrix.mixing_matrix}
\alias{as.data.frame.mixing_matrix}
\title{Construct a graph object's mixing matrix.}
\usage{
rep_as_mixing_matrix(x, vrt_attr, drop_loops = FALSE,
  leave_raw = FALSE)

is_mixing_matrix(x)

\method{print}{mixing_matrix}(mixing_matrix, ...)

\method{as.matrix}{mixing_matrix}(mixing_matrix)

\method{as.data.frame}{mixing_matrix}(mixing_matrix, ...)
}
\arguments{
\item{x}{An \code{bridge_net}, \code{igraph}, \code{network}, \code{tbl_graph}, or \code{edgelist} object.}

\item{vrt_attr}{\code{character} indicating which vertex attribute to use.}

\item{drop_loops}{\code{logical} (default: \code{FALSE}) indicating whether to drop loop edges
prior to calculations.}

\item{leave_raw}{\code{logical} (default: \code{FALSE}) indicating whether the returned object
should be a raw \code{matrix} rather than a \code{mixing_matrix} object.}

\item{mixing_matrix}{An object of class \code{mixing_matrix}.}

\item{...}{Additional arguments to be passed to or used by other methods.}
}
\value{
\code{mixing_matrix} or \code{matrix}
}
\description{
Using a categorical vertex attribute, construct a matrix depicting network mixing
patterns.
}
\examples{
library(snatools)

data("sampson", package = "ergm")

samplike \%>\% rep_as_mixing_matrix(vrt_attr = "group")

samplike \%>\% rep_as_mixing_matrix(vrt_attr = "group", leave_raw = TRUE)

}
\seealso{
\code{\link[network:mixingmatrix]{network::mixingmatrix()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
