% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_edgelist}
\alias{rep_edgelist}
\alias{rep_edgelist.igraph}
\alias{rep_edgelist.network}
\alias{rep_attr_el}
\alias{rep_attr_el.attr_adj_mat}
\alias{rep_attr_el.igraph}
\alias{rep_attr_el.network}
\title{Build a networks's edge list representation.}
\usage{
rep_edgelist(x, ...)

\method{rep_edgelist}{igraph}(ig, use_names = FALSE)

\method{rep_edgelist}{network}(x, use_names = FALSE)

rep_attr_el(x, vrt_attr)

\method{rep_attr_el}{attr_adj_mat}(x, vrt_attr)

\method{rep_attr_el}{igraph}(x, vrt_attr)

\method{rep_attr_el}{network}(x, vrt_attr)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{use_names}{\code{logical} Whether the returned edge list should use vertex names instead of their
indices. \cr
Default: \code{FALSE}}

\item{vrt_attr}{\code{character}. Name of target vertex attribute.}
}
\value{
\code{matrix} \cr
\itemize{
\item \code{mode()}:
\itemize{
\item \code{numeric} if \code{use_names} is \code{FALSE}.
\item \code{character} if \code{use_names} is \code{TRUE}.
}
\item \code{dim()}
\itemize{
\item \code{igraph::ecount(x)} by 2 if \code{x} is an \code{igraph} object.
\item \code{network::network.edgecount(x)} by 2 if \code{x} is a \code{network} object.
}
}
}
\description{
\code{rep_edgelist()} creates edge lists in a \emph{consistent} format across classes.
}
\details{
\itemize{
\item If \code{x} is \emph{directed}, columns are named \code{"from"} and \code{"to"} to indicate the
tie's direction.
\item If \code{x} is \emph{undirected}, columns are named \code{"vert1"} and \code{"vert2"}.
}

\code{rep_attr_el()} creates an edge list replacing vertex names with the category
in \code{vrt_attr} to which each vertex belongs.
}
\examples{
library(snatools)

# igraph ==============================================================================
data("karate", package = "igraphdata") # loads "Zachary's Karate Club" as `karate`
karate

karate \%>\% 
  rep_edgelist() \%>\% 
  head()
  
karate \%>\% 
  rep_edgelist(use_names = TRUE) \%>\% 
  head()

# network =============================================================================
data("sampson", package = "ergm") # loads "Sampson's Monks" as `samplike`
samplike

samplike \%>\% 
  rep_edgelist() \%>\% 
  head()
  
samplike \%>\% 
  rep_edgelist(use_names = TRUE) \%>\% 
  head()

}
\seealso{
\code{\link[igraph:as_edgelist]{igraph::as_edgelist()}}, \code{\link[network:as.edgelist]{network::as.edgelist()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
