% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_as_edgelist.R
\name{sna_as_edgelist}
\alias{sna_as_edgelist}
\alias{sna_as_edgelist.igraph}
\alias{sna_as_edgelist.network}
\title{Build an edge list of a graph}
\usage{
sna_as_edgelist(x, use_names = FALSE)

\method{sna_as_edgelist}{igraph}(ig, use_names = FALSE)

\method{sna_as_edgelist}{network}(nw, use_names = FALSE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}
}
\value{
A 2 column \code{matrix}. If graph is directed, the first column is always
the source of a tie.
}
\description{
Build an edge list of a graph
}
