% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna_as_edgelist.R
\name{sna_get_edgelist}
\alias{sna_get_edgelist}
\alias{sna_get_edgelist.igraph}
\alias{sna_get_edgelist.network}
\title{Build an edge list of a graph}
\usage{
sna_get_edgelist(x, use_names = FALSE)

\method{sna_get_edgelist}{igraph}(ig, use_names = FALSE)

\method{sna_get_edgelist}{network}(nw, use_names = FALSE)
}
\arguments{
\item{x}{An \code{igraph} or \code{network} object.}

\item{use_names}{\code{logical} Whether the returned edge list should use vertex names
instead of their indices. \cr
Default: \code{FALSE}}
}
\value{
A 2 column \code{matrix}. If graph is directed, the first column is always
the source of a tie.
}
\description{
Build an edge list of a graph
}
\examples{
library(snatools)

# igraph ==============================================================================
data("karate", package = "igraphdata")

karate \%>\% 
  sna_get_edgelist() \%>\% 
  head()
  
karate \%>\% 
  sna_get_edgelist(use_names = TRUE) \%>\% 
  head()

# network =============================================================================
data("sampson", package = "ergm")

samplike \%>\% 
  sna_get_edgelist() \%>\% 
  head()
  
samplike \%>\% 
  sna_get_edgelist(use_names = TRUE) \%>\% 
  head()

}
\seealso{
\code{\link[igraph:as_edgelist]{igraph::as_edgelist()}}, \code{\link[network:as.edgelist]{network::as.edgelist()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
