% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snatools-package.R
\docType{package}
\name{snatools}
\alias{snatools}
\alias{snatools-package}
\title{snatools: An R toolkit to bridge graph classes and streamline network analytic workflows.}
\description{
What the package does (one paragraph).
}
\author{
Brendan Knapp <brendan.g.knapp@gmail.com>
}
