% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_attr_names}
\alias{vrt_attr_names}
\alias{vrt_attr_names.sna_net}
\alias{vrt_attr_names.igraph}
\alias{vrt_attr_names.network}
\alias{vrt_attr_names.tbl_graph}
\title{Extract a graph objects's vertex attribute names.}
\usage{
vrt_attr_names(x)

\method{vrt_attr_names}{sna_net}(x)

\method{vrt_attr_names}{igraph}(x)

\method{vrt_attr_names}{network}(x)

\method{vrt_attr_names}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{sna_net}, \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{character} \code{vector} listing the names of \code{x}'s vertex attributes.
}
\description{
Extract a graph objects's vertex attribute names.
}
\examples{
library(snatools)

adj_mat <- rbinom(16L, 1L, 0.3) \%>\% 
  matrix(nrow = 4L, dimnames = list(letters[1:4], letters[1:4]))

ig <- adj_mat \%>\% 
  igraph::graph_from_adjacency_matrix() \%>\% 
  igraph::set_vertex_attr("another_attr", value = sample(seq_len(igraph::vcount(.)))) 

vrt_attr_names(ig)

nw <- network::as.network(adj_mat)
network::set.vertex.attribute(nw, "another_attr", value = sample(seq_len(network::network.size(nw))))

vrt_attr_names(nw)

tbl_g <- tidygraph::as_tbl_graph(ig)

vrt_attr_names(tbl_g)

sna_g <- as_sna_net(ig)

vrt_attr_names(sna_g)

}
\seealso{
\code{\link[=vrt_get_attr]{vrt_get_attr()}}, \code{\link[=vrt_get_attr_df]{vrt_get_attr_df()}}, \code{\link[igraph:vertex_attr_names]{igraph::vertex_attr_names()}},
\code{\link[network:list.vertex.attributes]{network::list.vertex.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
