% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_attr_names}
\alias{vrt_attr_names}
\alias{vrt_attr_names.igraph}
\alias{vrt_attr_names.network}
\alias{vrt_attr_names.tbl_graph}
\title{Get Vertex Attribute Names}
\usage{
vrt_attr_names(x)

\method{vrt_attr_names}{igraph}(x)

\method{vrt_attr_names}{network}(x)

\method{vrt_attr_names}{tbl_graph}(x)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}
}
\value{
\code{character} \code{vector} of \code{x}'s vertex attribute names or  \code{NULL} if
no vertex attributes exist.
}
\description{
Get Vertex Attribute Names
}
\seealso{
\code{\link[igraph:vertex_attr_names]{igraph::vertex_attr_names()}}

\code{\link[network:list.vertex.attributes]{network::list.vertex.attributes()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
