% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-attributes.R
\name{vrt_get_attr}
\alias{vrt_get_attr}
\alias{vrt_get_attr.igraph}
\alias{vrt_get_attr.network}
\alias{vrt_get_attr.tbl_graph}
\title{Get a Vertex Attribute}
\usage{
vrt_get_attr(x, vrt_attr)

\method{vrt_get_attr}{igraph}(x, vrt_attr)

\method{vrt_get_attr}{network}(x, vrt_attr)

\method{vrt_get_attr}{tbl_graph}(x, vrt_attr)
}
\arguments{
\item{x}{An \code{igraph}, \code{network}, or \code{tbl_graph}.}

\item{vrt_attr}{\code{character} specifying the target vertex attribute.}
}
\value{
\code{vector} containing the vertex attributes specified by \code{vrt_attr}.
}
\description{
Get a Vertex Attribute
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
